% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addText.R
\name{addText}
\alias{addText}
\title{Image Text Incrustation}
\usage{
addText(image, text, color, xoff = 0, yoff = 0, corner = "TL")
}
\arguments{
\item{image}{a [0,1] image.}

\item{text}{a character string.}

\item{color}{a character string. color of the text.}

\item{xoff}{positive integer. x offset in image to start writing text.}

\item{yoff}{positive integer. y offset in image to start writing text.}

\item{corner}{a character string. where to position text in the image. Allowed are "TL", "TR", "BL", "BR", for top-left, top-right, bottom-left, bottom-right, respectively.}
}
\value{
an image with text added.
}
\description{
Adds Text to image.
}
\details{
One-lined text will be added so has to be fully contained within image and anchored at desired corner plus x and y offset from it.
}
\keyword{internal}
