% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addText.R
\name{addText2}
\alias{addText2}
\title{Image Text Incrustation}
\usage{
addText2(
  image,
  text,
  color,
  xoff = 0,
  yoff = 0,
  anchor = c(1, 0),
  vjust = c("T", "M", "B")[1],
  hjust = c("L", "C", "R")[1]
)
}
\arguments{
\item{image}{a [0,1] image.}

\item{text}{a character string.}

\item{color}{a character string. color of the text.}

\item{xoff}{x offset in image to start writing text according to its anchorage and justification. Default is \code{0}.}

\item{yoff}{y offset in image to start writing text according to its anchorage and justification. Default is \code{0}.}

\item{anchor}{numeric vector of text anchorage in normalized coordinates starting from bottom-left. Default is \code{c(1,0)}, for top-left anchorage.}

\item{vjust}{vertical justification of the text. Default is \code{"T"}, allowed is one of \code{"T"}, \code{"M"}, and \code{"B"} for top, middle or bottom justification, respectively.}

\item{hjust}{horizontal justification of the text. Default is \code{"L"}, allowed is one of \code{"L"}, \code{"C"}, and \code{"R"} for left, center or right justification, respectively.}
}
\value{
an image with text added.
}
\description{
Adds Text to image.
}
\details{
One-lined text will be added so has to be fully contained within image.
}
\keyword{internal}
