% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_utils.R
\name{applyTrans}
\alias{applyTrans}
\title{Apply Transformation}
\usage{
applyTrans(x, trans, inverse = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{trans}{the object returned by parseTrans().}

\item{inverse}{whether or not to apply the inverse transformation. Default is FALSE.}
}
\value{
the transformation of the input
}
\description{
Helper apply transformation.
}
\details{
for the moment, in addition to no transformation, only "smoothLinLog", "asinh", and "sqrt" are supported.
}
\keyword{internal}
