% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert}
\alias{assert}
\title{Assert that Certain Conditions are Met}
\usage{
assert(x, len = NULL, cla = NULL, typ = NULL, alw = NULL, fun = "stop")
}
\arguments{
\item{x}{variable to test}

\item{len}{integer vector of allowed length for x. Default is NULL, for not checking this parameter.}

\item{cla}{character vector of allowed classes of x. Default is NULL, for not checking this parameter.}

\item{typ}{character vector of allowed types of x. Default is NULL, for not checking this parameter.}

\item{alw}{allowed values for x. Default is NULL, for not checking this parameter.}

\item{fun}{function to execute when mandatory parameters are not met. Default is "stop". Allowed are "stop","warning","message","return".}
}
\description{
Ensures that a variable respects several parameters
}
\details{
/!\ alw parameter when used should be coercible to a logical, integer, numeric, complex or character vector. Otherwise, an error will be thrown.
}
\keyword{internal}
