% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_closure.R
\name{avg_closure}
\alias{avg_closure}
\title{Finds the Average Length of IFC Moratoria.}
\usage{
avg_closure(data, university, start1, end1, start2, end2)
}
\arguments{
\item{data}{an object/dataframe/tibble.}

\item{university}{the column in the object that contains the universities.}

\item{start1}{the column in the object that contains the start date of the first moratorium.}

\item{end1}{the column in the object that contains the  end date of the first moratorium.}

\item{start2}{the column in the object that contains the start date of the second moratorium.}

\item{end2}{the column in the object that contains the end date of the second moratorium.}
}
\description{
\code{avg_closure} returns the average length of fraternity moratoria in days.
}
\details{
This function relies on dplyr, lubridate, and rlang to work. Moreover, this function
does a few operations. First, it selects only the columns of interest, then, it splits
a tibble into two: one for first set of closures, and one for second closures (third coming).
Next, it creates columns with the length of each of these closures. Finally, it binds together
the two tibbles, and takes the mean of the length.
}
\examples{
avg_closure(data, university, start1, end1, start2, end2)

}
