% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base64.R
\name{base64_decode}
\alias{base64_decode}
\title{Base64 to Raw Decoding}
\source{
\url{https://en.wikibooks.org/wiki/Algorithm_Implementation/Miscellaneous/Base64}
}
\usage{
base64_decode(x, url = identical(attr(x, "url", exact = TRUE), TRUE))
}
\arguments{
\item{x}{a base64 encoded string.}

\item{url}{a bool whether to decode from base64 url encoding. Default is \code{identical(attr(x, "url", exact = TRUE), TRUE)}.}
}
\description{
Decodes base64 string to raw vector
}
\keyword{internal}
