% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base64.R
\name{base64_encode}
\alias{base64_encode}
\title{Raw to Base64 Encoding}
\source{
\url{https://en.wikibooks.org/wiki/Algorithm_Implementation/Miscellaneous/Base64}
}
\usage{
base64_encode(x, url = FALSE)
}
\arguments{
\item{x}{a raw vector.}

\item{url}{a bool whether to encode for base64 url string. Default is \code{FALSE}.}
}
\description{
Encodes raw vector to base64 string
}
\keyword{internal}
