% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{base_hist_constr}
\alias{base_hist_constr}
\title{Histogram Constructor for 'base' Plot}
\usage{
base_hist_constr(
  x,
  type,
  br,
  normalize,
  fill,
  smooth,
  lwd,
  lty,
  col,
  alpha,
  border,
  include.lowest = TRUE,
  right = TRUE
)
}
\arguments{
\item{x}{a vector of values for which the histogram is desired.}

\item{type}{histogram type. Default is missing. Allowed are "count" and "percent".}

\item{br}{breakpoints given an interval and the number of pieces to break it into.}

\item{normalize}{whether to normalize. Default is missing.}

\item{fill}{whether to fill. Default is missing.}

\item{smooth}{whether to smooth. Default is missing.}

\item{lwd, lty, col, alpha, border}{graphical parameters. See par() from package 'graphics'.}

\item{include.lowest}{logical; if TRUE, an x[i] equal to the breaks value will be included in the first (or last, for right = FALSE) bar. This will be ignored (with a warning) unless breaks is a vector.}

\item{right}{logical; if TRUE, the histogram cells are right-closed (left open) intervals.}
}
\description{
Helper to create histogram.
}
\keyword{internal}
