% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{base_theme}
\alias{base_theme}
\title{`IFC_plot` Base Theme Creation}
\usage{
base_theme(obj = list(), color_mode = c("white", "black")[1], lt = NULL)
}
\arguments{
\item{obj}{an object of class `IFC_plot` as created by \code{\link{plotGraph}}.}

\item{color_mode}{whether to extract colors from 'obj' in white or black mode. Default is "white".}

\item{lt}{an existing lattice theme to modify according to \code{'color_mode'}.}
}
\value{
a list of graphical parameters for 'base' and 'raster' backend.
}
\description{
Helper to create base theme from `IFC_plot`
}
\keyword{internal}
