% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{base_title}
\alias{base_title}
\title{`IFC_plot` Base Title Annotation}
\usage{
base_title(obj)
}
\arguments{
\item{obj}{an object of class `IFC_plot` as created by \code{\link{plotGraph}}.}
}
\value{
nothing, annotate plot with 'base' and 'raster' backend.
}
\description{
Helper to add title (and subtitle) when drawing `IFC_plot`
}
\keyword{internal}
