% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{base_xaxis}
\alias{base_xaxis}
\title{`IFC_plot` Base x Axis Annotation}
\usage{
base_xaxis(obj)
}
\arguments{
\item{obj}{an object of class `IFC_plot` as created by \code{\link{plotGraph}}.}
}
\value{
nothing, annotate plot with 'base' and 'raster' backend.
}
\description{
Helper to add x axis when drawing `IFC_plot`
}
\keyword{internal}
