% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{base_yaxis}
\alias{base_yaxis}
\title{`IFC_plot` Base y Axis Annotation}
\usage{
base_yaxis(obj)
}
\arguments{
\item{obj}{an object of class `IFC_plot` as created by \code{\link{plotGraph}}.}
}
\value{
nothing, annotate plot with 'base' and 'raster' backend.
}
\description{
Helper to add y axis when drawing `IFC_plot`
}
\keyword{internal}
