% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{BatchStatsReport}
\alias{BatchStatsReport}
\title{Batch Generation of Statistic Report}
\usage{
BatchStatsReport(fileName, obj, stats, gating, display_progress = FALSE, ...)
}
\arguments{
\item{fileName, obj}{either one or the other. Path to file(s) to read from for 'fileName' or list of `IFC_data` objects for obj.}

\item{stats}{template defining stats to compute as extracted by getSTATSREPORT().}

\item{gating}{an `IFC_gating` object as extracted by readGatingStrategy(). Default is missing.
If not missing, each `IFC_data` provided in 'obj' or read from 'fileName' will be passed to applyGatingStrategy() before creating the report.}

\item{display_progress}{whether to display a progress bar. Default is TRUE.}

\item{...}{other parameters to be passed.}
}
\value{
a data.frame of statistics
}
\description{
Generates statistics report on batch of files or `IFC_data` objects.
}
\keyword{internal}
