% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFeature.R
\name{buildFeature}
\alias{buildFeature}
\title{IFC Feature Coercion}
\usage{
buildFeature(
  name,
  type = c("single", "combined", "computed")[1],
  def = "Camera Line Number",
  val = NULL,
  ...
)
}
\arguments{
\item{name}{feature's name. If missing, it will be determined thanks to def.}

\item{type}{feature's type. Default is "single". Allowed are "single", "combined", "computed".}

\item{def}{definition of the feature. Default is "Camera Line Number".}

\item{val}{a coercible to numeric vector of feature values. Default is NULL.\cr
Note that although not mandatory for \code{\link{buildFeature}}it has to be provided to
allow feature export in \code{\link{ExportToDAF}} and \code{\link{data_add_features}}.}

\item{...}{Other arguments to be passed.}
}
\value{
a list containing all feature information.
}
\description{
Helper to build a list to allow feature export.
}
