% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiff_utils.R
\name{buildIFD}
\alias{buildIFD}
\title{Image Field Directory Builder}
\usage{
buildIFD(val, typ, tag, endianness = .Platform$endian)
}
\arguments{
\item{val}{the value of the ifd}

\item{typ}{desired ifd type}

\item{tag}{the desired ifd 'tag'}

\item{endianness}{the desired endian-ness ("big" or "little"). Default is .Platform$endian.}
}
\value{
NULL or a list of 2 members:\cr
-min_content: the minimal ifd content,\cr
-add_content: the additional ifd content if 'val' converted to raw does not fit in 4 bytes.
}
\description{
build IFD
}
\details{
if 'val' if of type "character", 'tag' is automatically set to 2.\cr
if 'val' is of length 0 NULL is returned.
}
\keyword{internal}
