% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redefine_utils.R
\name{buildMask}
\alias{buildMask}
\title{IFC Mask Coercion}
\usage{
buildMask(
  masks,
  images,
  definition = TRUE,
  to_match_mask = NULL,
  to_replace_mask = NULL,
  to_match_image = NULL,
  to_replace_image = NULL,
  ...
)
}
\arguments{
\item{masks}{an `IFC_masks` object or a data.frame containing masks definition and name. Default is missing.}

\item{images}{a data.frame containing images definition. Default is missing.}

\item{definition}{whether to compute masks definition or masks names. Default is TRUE, to compute definition.}

\item{to_match_mask}{a vector of masks names to use for matching 'masks' names. Default is NULL}

\item{to_replace_mask}{a vector of masks names to use for replacing 'masks' names. Default is NULL}

\item{to_match_image}{a vector of images names to use for matching 'images' names. Default is NULL}

\item{to_replace_image}{a vector of images names to use for replacing 'images' names. Default is NULL}

\item{...}{Other arguments to be passed.}
}
\value{
a vector of masks definition or names depending on 'definition' parameter.
}
\description{
Helper to build a mask.
}
\details{
function that can either change definition or name
it will be used in a loop to incorporate new definition.
causing name to be changed. allowing new redefinition of mask 
dependent on other mask to be changed, and so on
when a mask or an image name is not found because not yet defined 
an error is raised and catched
TODO maybe this error catching causes some overhead to be improved
}
\keyword{internal}
