% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRegion.R
\name{buildRegion}
\alias{buildRegion}
\title{IFC Region Coercion}
\usage{
buildRegion(
  type,
  label,
  cx,
  cy,
  color,
  lightcolor,
  ismarker = "false",
  doesnotoverride = "false",
  xlogrange,
  ylogrange,
  x,
  y,
  ...
)
}
\arguments{
\item{type}{Region's type. Either "line", "rect", "poly" or "oval".}

\item{label}{label of the region.}

\item{cx}{x label's position. If not provided x center will be used.}

\item{cy}{y label's position. If not provided y center will be used.}

\item{color}{color of the region. See \code{\link{paletteIFC}} for allowed colors.}

\item{lightcolor}{lightcolor of the region. See \code{\link{paletteIFC}} for allowed colors.}

\item{ismarker}{Default is 'false'. Allowed are 'true' or 'false'. Used for compatibility with amnis file but role remains unknown.}

\item{doesnotoverride}{Default is 'false'. Allowed are 'true' or 'false'. Used for compatibility with amnis file but role remains unknown.}

\item{xlogrange}{determines hyper parameter of smoothLinLog transformation for x-axis. Default is "P" for no transformation.}

\item{ylogrange}{determines hyper parameter of smoothLinLog transformation for y-axis. Default is "P" for no transformation.}

\item{x}{vector of x vertices values.}

\item{y}{vector of y vertices values.}

\item{...}{Other arguments to be passed.}
}
\value{
a list containing all region information.
}
\description{
Helper to build a list to allow region export.
}
