% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{calcDensity}
\alias{calcDensity}
\title{2D Binned Kernel Density Estimation}
\source{
copyed from \pkg{grDevices}
}
\usage{
calcDensity(x, nbin, bandwidth, range.x)
}
\description{
Helper to compute density plot
}
\author{
R Core Team, Florian Hahne at FHCRC, originally
}
\keyword{internal}
