% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkColor.R
\name{checkColor}
\alias{checkColor}
\title{Character Color Check and RGB Conversion}
\usage{
checkColor(color)
}
\arguments{
\item{color}{a character string.}
}
\description{
Checks that color is R compatible by converting it to RGB matrix.
}
\keyword{internal}
