% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkObj.R
\name{checkObj}
\alias{checkObj}
\title{`IFC_data` Object Check}
\usage{
checkObj(obj)
}
\arguments{
\item{obj}{an `IFC_data` object.}
}
\value{
an `IFC_data` object.
}
\description{
Ensures `IFC_data` obj$features correctly reflects actual number of collected objects.
When 'obj' is from rif or cif, number of objects can be different from actual number of collected objects
e.g. when it comes from merged of subset of file. There is no way to link an object with its feature value,
so the only solution is to remove all features and only keep "Object Number".
}
\keyword{internal}
