% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportToReport.R
\name{CheckReportFileCreation}
\alias{CheckReportFileCreation}
\title{Checker for Report File Creation}
\usage{
CheckReportFileCreation(fileName, write_to, overwrite = FALSE)
}
\arguments{
\item{fileName}{default fileName path.}

\item{write_to}{pattern used to export file(s).
Placeholders, like c("\%d/\%s_fromR.pdf", "\%d/\%s_fromR.csv"), will be substituted:\cr
-\%d: with full path directory of 'fileName'\cr
-\%p: with first parent directory of 'fileName'\cr
-\%e: with extension of 'fileName' (without leading .)\cr
-\%s: with shortname from 'fileName' (i.e. basename without extension).\cr
Exported file(s) extension(s) will be deduced from this pattern. Note that has to be a .pdf and/or .csv.}

\item{overwrite}{whether to overwrite file or not. Default is FALSE.
Note that if TRUE, it will overwrite file. In addition a warning message will be sent.}
}
\value{
a list with path(s) to pdf and csv file and overwritten status
}
\description{
Checks if report files can be created.
}
\keyword{internal}
