% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checksumIFC.R
\name{checksumIFC}
\alias{checksumIFC}
\title{IFC Files Checksum}
\usage{
checksumIFC(fileName, ...)
}
\arguments{
\item{fileName}{path to file.}

\item{...}{arguments to pass to \code{\link{checksumDAF}} or \code{\link{checksumXIF}}.}
}
\value{
an integer corresponding to IFC file checksum.
}
\description{
This function returns RIF/CIF checksum.
Checksum is the sum of img IFDs (Image Field Directory) offsets of objects 0, 1, 2, 3 and 4.
}
\details{
if fileName is a DAF file, then CIF checksum is computed from images values found in DAF.
}
