% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checksumXIF.R
\name{checksumXIF}
\alias{checksumXIF}
\title{RIF/CIF File Checksum}
\usage{
checksumXIF(fileName, ...)
}
\arguments{
\item{fileName}{path to file.}

\item{...}{other arguments to be passed.}
}
\description{
This function returns RIF/CIF checksum.
Checksum is the sum of img IFDs (Image Field Directory) offsets of objects 0, 1, 2, 3 and 4.
}
\keyword{internal}
