% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiff_utils.R
\name{collapse_raw}
\alias{collapse_raw}
\title{Raw Vectors Collapse}
\usage{
collapse_raw(x, collapse = as.raw(124))
}
\arguments{
\item{x}{a list of raw vectors.}
}
\value{
a collapsed raw vector
}
\description{
Collapses raw vectors together
}
\keyword{internal}
