% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compensation.R
\name{CompensateFromCIF}
\alias{CompensateFromCIF}
\title{Shorcut for Batch CIF Images Compensation}
\usage{
CompensateFromCIF(
  fileName,
  objects,
  offsets,
  spillover,
  display_progress = TRUE,
  ...
)
}
\arguments{
\item{objects}{integer vector, IDEAS objects ids numbers to use.
This argument is not mandatory, if missing, the default, all objects will be used.}

\item{offsets}{object of class `IFC_offset`. 
This argument is not mandatory but it may allow to save time for repeated image export on same file.}

\item{spillover}{spillover matrix to apply. Default is missing to do nothing.
When set to NULL images will be decompensated using spillover matrix set within .cif file. Otherwise,
the images will be decompensated with spillover matrix set within the file and recompensated with the one provided.}

\item{display_progress}{whether to display a progress bar. Default is TRUE.}

\item{...}{other arguments to be passed.}
}
\value{
A list of matrices/arrays of images corresponding to objects extracted.
}
\description{
Function to shortcut extraction and CIF file compensation / decompensation
}
\details{
this function is experimental inputs and outputs may change in the future.
}
\keyword{internal}
