% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compensation.R
\name{CompensateFromRIF}
\alias{CompensateFromRIF}
\title{Shorcut for Batch RIF Images Spatial Correction and Compensation}
\usage{
CompensateFromRIF(..., objects, offsets, spillover, display_progress = TRUE)
}
\arguments{
\item{...}{arguments to be passed to \code{\link{objectExtract}} with the exception of 'ifd' and 'bypass'(=TRUE).\cr
If 'param' is provided 'export'(="matrix"), 'mode'(="raw"), 'removal'(="none"),
'composite'(=""), 'selection'(="all"), 'size'(=c(0, 0)), 'force_width'(=FALSE), 'spatial_correction'(=TRUE), will be overwritten.\cr
If 'offsets' are not provided extra arguments can also be passed with ... \code{\link{getOffsets}}.\cr
/!\ If not any of 'fileName', 'info' and 'param' can be found in ... then attr(offsets, "fileName_image") will be used as 'fileName' input parameter to pass to \code{\link{objectParam}}.}

\item{objects}{integer vector, IDEAS objects ids numbers to use.
This argument is not mandatory, if missing, the default, all objects will be used.}

\item{offsets}{object of class `IFC_offset`. 
This argument is not mandatory but it may allow to save time for repeated image export on same file.}

\item{spillover}{spillover matrix to apply. Default is missing to apply no compensation.
In such case, you will get images with spatial correction only. However, when set to NULL images 
will be compensated using spillover matrix set during acquisition. If not NULL the provided spillover 
matrix will be used to compensate images.}

\item{display_progress}{whether to display a progress bar. Default is TRUE.}
}
\value{
A list of matrices/arrays of images corresponding to objects extracted.
}
\description{
Function to shortcut extraction and RIF file spatial correction and compensation
}
\details{
this function is experimental inputs and outputs may change in the future.
}
\keyword{internal}
