% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs.R
\name{convert_spillover}
\alias{convert_spillover}
\title{Spillover Converter}
\usage{
convert_spillover(spillover)
}
\arguments{
\item{spillover}{either a spillover matrix or a spillover keyword}
}
\value{
if 'spillover' is a matrix, it returns a string. If 'spillover' is a string, it returns a matrix. In all cases if spillover is of length 0, it will return NULL.
}
\description{
Converts spillover matrix to spillover keyword and reversely
}
\keyword{internal}
