% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{convert_to_baseplot}
\alias{convert_to_baseplot}
\title{IFC Graph Conversion to 'base' Plot}
\usage{
convert_to_baseplot(obj)
}
\arguments{
\item{obj}{an object of class `IFC_plot` as created by \code{\link{plotGraph}}.}
}
\description{
Helper to convert `IFC_plot` to 'base' plot.
}
\keyword{internal}
