% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{coord_to_px}
\alias{coord_to_px}
\title{User's Coordinates to Pixels Conversion}
\usage{
coord_to_px(coord, coordmap, clipedge = FALSE)
}
\arguments{
\item{coord}{coordinates in user system. A matrix where rows are points and with at least 2 columns named "x" and "y" for x and y coordinates, respectively.}

\item{coordmap}{current device adjusted coordinates. Default is missing.}

\item{clipedge}{whether to clip points outside of plotting region to the edge. Default is FALSE.}
}
\value{
a 2-columns matrix with "x" and "y" coordinates.
}
\description{
Helper map user's coordinates to pixels
}
\keyword{internal}
