% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_align}
\alias{cpp_align}
\title{Spatial Offsets Image Correction}
\arguments{
\item{mat, }{a NumericMatrix.}

\item{dx, }{a double x spatial offset. Default is NA_REAL for no change.}

\item{dy, }{a double y spatial offset. Default is NA_REAL for no change.}

\item{add_noise}{logical, if true adds normal noise when at least one new dimension is larger than original mat dimensions. Rcpp::rnorm() function is used. Default is true.}

\item{bg}{double, mean value of the background added if add_noise is true. Default is 0.0.}

\item{sd}{double, standard deviation of the background added if add_noise is true. Default is 0.0.}
}
\value{
a NumericMatrix.
}
\description{
This function uses bilinear interpolation to apply spatial offset correction on image
}
\details{
It is intended to be applied on raw images matrices from .rif files so has to generate spatial offset corrected image matrices.\cr
See William E. Ortyn et al. Sensitivity Measurement and Compensation in Spectral Imaging. Cytometry A 69 852-862 (2006).
\doi{10.1002/cyto.a.20306}
}
\keyword{internal}
