% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_AND_M}
\alias{cpp_AND_M}
\title{Matrix List And Logic}
\arguments{
\item{list}{a list of logical matrices.}
}
\value{
a logical matrix.
}
\description{
This function takes a list of matrices and returns the AND operation applied on these matrices.
}
\keyword{internal}
