% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_as_nativeRaster}
\alias{cpp_as_nativeRaster}
\title{Image to Native Raster Conversion}
\arguments{
\item{x}{an IntegerVecter /!\ It should be coercible to 3D array [height, width, rgba]}
}
\value{
a nativeRaster IntegerMatrix
}
\description{
Converts 3D image array to nativeRaster
}
\keyword{internal}
