% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_assert}
\alias{cpp_assert}
\title{Input Parameters Assertive Tool}
\arguments{
\item{len}{IntegerVector, of allowed length for x. Default is NULL, for not checking this parameter.}

\item{cla}{CharacterVector, of allowed classes of x. Default is NULL, for not checking this parameter.}

\item{typ}{CharacterVector, of allowed types of x. Default is NULL, for not checking this parameter.}

\item{Robject}{of allowed values for x (will be passed to cpp_allowed). Default is NULL, for not checking this parameter.}

\item{fun}{CharacterVector, function to execute when mandatory parameters are not met. Default is "stop". Allowed are "stop","warning","message","return".
fun is placed in cpp_assert() in order to check it is correct before being used in assert() function.}
}
\description{
Ensures that x respects several parameters
}
\keyword{internal}
