% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_base64_decode}
\alias{cpp_base64_decode}
\title{Base64 to Raw Conversion}
\arguments{
\item{x}{a string.}

\item{url}{a bool, whether to convert for url. Default is false.}
}
\value{
a RawVector, representing the decoding of base64 string.
}
\description{
Converts a base64 string to raw vector.
}
\keyword{internal}
