% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_base64_encode}
\alias{cpp_base64_encode}
\title{Raw to Base64 Conversion}
\arguments{
\item{x}{RawVector.}
}
\value{
a string, representing the base64 encoding of x.
}
\description{
Converts a raw vector to base64 string.
}
\keyword{internal}
