% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_cast_image}
\alias{cpp_cast_image}
\title{Cast Image}
\arguments{
\item{x}{SEXP, the image to cast}

\item{what}{uint8_t, type to use for casting. Default is \code{1}.
Allowed are 1=uint8_t, 2=int8_t, 3=uint16_t, 4=int16_t, 5=uint32_t, 6=int32_t, 7=float_t, 8=double_t.}

\item{swap}{bool, whether single scalar values of \code{x} should be swap. Default is \code{false}.}
}
\value{
a Rcpp::RawVector
}
\description{
Casts image from RAW, INT or REAL SEXP vector
}
\keyword{internal}
