% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_checksum}
\alias{cpp_checksum}
\title{Checksum for RIF/CIF}
\source{
TIFF 6.0 specifications available at \url{https://www.adobe.io/open/standards/TIFF.html}
}
\arguments{
\item{fname}{string, path to file.}
}
\value{
an integer vector with offsets of IFDs found.
}
\description{
Computes sum of img IFDs (Image Field Directory) offsets of objects 0, 1, 2, 3 and 4.
}
\keyword{internal}
