% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_checkTIFF}
\alias{cpp_checkTIFF}
\title{TIFF Checker}
\source{
TIFF 6.0 specifications available at \url{https://www.adobe.io/open/standards/TIFF.html}
}
\arguments{
\item{fname}{string, path to file.}
}
\description{
Checks if file is a TIFF.
}
\details{
If file is a TIFF it returns endianness of file, 'big' or 'little.
Otherwise, it shows an error and returns an empty string.
}
\keyword{internal}
