% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_cleanse}
\alias{cpp_cleanse}
\title{Matrix Cleanser}
\arguments{
\item{mat}{a NumericMatrix.}

\item{msk}{a IntegerMatrix.}

\item{add_noise}{bool, whether to add normal noise or not, \code{Rcpp::Rf_rnorm(bg, sd)} function is used. Default is \code{true}.}

\item{bg}{double, mean value of the background added if \code{'add_noise'} is \code{true}. Default is \code{0.0}.}

\item{sd}{double, standard deviation of the background added if \code{'add_noise'} is \code{true}. Default is \code{0.0}.}
}
\value{
a NumericMatrix cleansed according to \code{'msk'}.
}
\description{
Replaces values in matrix mat according to mask msk.
Depending of \code{'add_noise'} parameter, values of \code{'mat'} will be replaced with noise or not.
}
\keyword{internal}
