% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_computeGamma}
\alias{cpp_computeGamma}
\title{Gamma Computation}
\arguments{
\item{V}{named NumericVector of channel display properties containing 'xmin', 'xmax', 'xmid' and 'ymid'.}
}
\description{
This function computes image gamma transformation value.
}
\keyword{internal}
