% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_coord_to_px}
\alias{cpp_coord_to_px}
\title{Coordinates to Pixels}
\arguments{
\item{x}{NumericVector of x-coordinates of the points.}

\item{y}{NumericVector of y-coordinates of the points.}

\item{param}{NumericVector of parameters to scale raw points coordinates to pixels coordinates.}
}
\value{
a 2 columns NumericMatrix of x and y pixels coordinates.
}
\description{
low-level function to compute pixels coordinates
}
\keyword{internal}
