% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_crop}
\alias{cpp_crop}
\title{Matrix Cropping}
\arguments{
\item{mat}{Matrix (Raw, Integer, Logical or Numeric).}

\item{new_height}{a R_len_t, giving the new height of returned mat. Default is 0 for no change. Negative values will remove rows from mat.}

\item{new_width}{a R_len_t, giving the new width of returned mat. Default is 0 for no change. Negative values will remove cols from mat.}

\item{front}{a bool, whether to apply cropping from front. Default is false.}
}
\value{
a cropped matrix.
}
\description{
Crops mat according to new_height and new_width parameters.
}
\keyword{internal}
