% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_ell_coord}
\alias{cpp_ell_coord}
\title{Ellipse Boundaries to Coordinates}
\arguments{
\item{bound_x}{NumericVector, x-boundaries of the ellipse.}

\item{bound_y}{NumericVector, y-boundaries of the ellipse.}
}
\description{
Transforms ellipse boundaries to usefull coordinates.
}
\keyword{internal}
