% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_fast_cbind_DF_DF}
\alias{cpp_fast_cbind_DF_DF}
\title{Dataframe and Dataframe Column Binding}
\arguments{
\item{Df1_}{a Nullable DataFrame.}

\item{Df2_}{a Nullable DataFrame.}

\item{add_id}{a bool determining if 1st column of returned object should be given 1 to nrow integers}
}
\value{
a DataFrame.
}
\description{
Combines numeric matrix by columns
}
\keyword{internal}
