% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_fast_cbind_DF_L}
\alias{cpp_fast_cbind_DF_L}
\title{Fast Dataframe and List Column Binding}
\arguments{
\item{Df_}{a Nullable DataFrame.}

\item{L_}{a Nullable List.}

\item{add_id}{a bool determining if 1st column of returned object should be given 1 to nrow integers}
}
\value{
a DataFrame.
}
\description{
Combines data.frame and list by columns
}
\keyword{internal}
