% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_fast_cbind_DF_M}
\alias{cpp_fast_cbind_DF_M}
\title{Fast Dataframe and Matrix Column Binding}
\arguments{
\item{Df_}{a Nullable DataFrame.}

\item{M_}{a Nullable NumericVector. /!\ But cast to NumericMatrix.}

\item{add_id}{a bool determining if 1st column of returned object should be given 1 to nrow integers}
}
\value{
a DataFrame.
}
\description{
Combines data.frame and matrix by columns
}
\keyword{internal}
