% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_fast_cbind_M_DF}
\alias{cpp_fast_cbind_M_DF}
\title{Fast Matrix and Dataframe Column Binding}
\arguments{
\item{M_}{a Nullable NumericVector. /!\ But cast to NumericMatrix.}

\item{Df_}{a Nullable DataFrame.}

\item{add_id}{a bool determining if 1st column of returned object should be given 1 to nrow integers}
}
\value{
a DataFrame.
}
\description{
Combines matrix and data.frame by columns
}
\keyword{internal}
