% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_fast_cbind_M_M}
\alias{cpp_fast_cbind_M_M}
\title{Matrix and Matrix Column Binding}
\arguments{
\item{M1_}{a Nullable NumericVector. /!\ But cast to NumericMatrix.}

\item{M2_}{a Nullable NumericVector. /!\ But cast to NumericMatrix.}

\item{add_id}{a bool determining if 1st column of returned object should be given 1 to nrow integers}
}
\value{
a NumericVector.
}
\description{
Combines numeric matrix by columns
}
\keyword{internal}
