% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_fast_factor}
\alias{cpp_fast_factor}
\title{Fast Factorize Vector}
\source{
adaptation from Kevin Ushey code \url{https://gallery.rcpp.org/articles/fast-factor-generation}
}
\arguments{
\item{x}{a SEXP only NILSXP, LGLSXP, INTSXP, REALSXP, STRSXP, RAWSXP are handled.}

\item{handleNA}{a bool specifying if NA should be returned as NA or if they should be attributed a unique integer.}
}
\value{
an IntegerVector, with attributes "levels" being the non-NA unique value(s) found
and "lvs" the total number of unique values found (including NA).
}
\description{
Makes factor with Rcpp
}
\details{
returned object should not be of class `factor` so has to prevent malformed factor' error when printing the result in R.
For this reason, it is aimed to be wrap in an R function to handle factor class / handleNA balance.\cr
e.g. either:
- structure(cpp_fast_factor(df, TRUE), class = "factor"), or,
- structure(cpp_fast_factor(unclass(df), FALSE), levels = NULL)
}
\keyword{internal}
