% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_fast_listAny}
\alias{cpp_fast_listAny}
\title{Use Rcpp to Apply Any on List members}
\arguments{
\item{L_}{a Nullable List.}
}
\value{
a LogicalVector.
}
\description{
Computes any across list members
}
\keyword{internal}
