% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_fast_rowAny}
\alias{cpp_fast_rowAny}
\title{Use Rcpp to Apply Any on Matrix Rows}
\arguments{
\item{M_}{a Nullable LogicalVector. /!\ But cast to LogicalMatrix}
}
\value{
a LogicalVector.
}
\description{
Computes any across matrix rows.
}
\keyword{internal}
