% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_fast_sample}
\alias{cpp_fast_sample}
\title{Use Rcpp for Sampling}
\arguments{
\item{n}{a R_len_t, max number integers to choose from.}

\item{size}{a R_len_t the desired size of return integers.}

\item{replace}{a bool determining if sampling should be done with replacement. Default is false.}
}
\description{
Create a sample of integers
}
\keyword{internal}
