% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_get_bytes_order}
\alias{cpp_get_bytes_order}
\title{Get Bytes Order}
\arguments{
\item{obj}{number of objects in the data.}

\item{byt_}{IntegerVector of number of bytes to take from 'ord_'.}

\item{ord_}{IntegerVector bytes order.}

\item{rev}{bool whether to reverse order or not. Default is false.}
}
\description{
This function expands bytes order to the whole data
}
\keyword{internal}
