% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_getoffsets_noid}
\alias{cpp_getoffsets_noid}
\title{IFC_offsets Computation without Id Determination}
\source{
TIFF 6.0 specifications available at \url{https://www.adobe.io/open/standards/TIFF.html}
}
\arguments{
\item{fname}{string, path to file.}

\item{obj_count}{R_len_t, numbers of objects present in the file. Default is 0.
If obj_count <= 0 then progress_bar is forced to false.}

\item{display_progress}{bool, whether to display a progress bar. Default is false.}

\item{verbose}{bool, whether to display information (use for debugging purpose). Default is false.}
}
\value{
an integer vector with offsets of IFDs found.
}
\description{
Returns offsets of the IFDs (Image Field Directory) within a TIFF file.
}
\keyword{internal}
