% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_getoffsets_wid}
\alias{cpp_getoffsets_wid}
\title{IFC_offsets Computation with Object Identification}
\source{
TIFF 6.0 specifications available at \url{https://www.adobe.io/open/standards/TIFF.html}
}
\arguments{
\item{fname}{string, path to file.}

\item{obj_count}{R_len_t, numbers of objects present in the file. Default is 0.
If obj_count <= 0 then progress_bar is forced to false.}

\item{display_progress}{bool, whether to display a progress bar. Default is false.}

\item{verbose}{bool, whether to display information (use for debugging purpose). Default is false.}
}
\value{
a list of integer vectors with OBJECT_ID, TYPE and OFFSET of IFDs found.
}
\description{
Returns offsets of the IFD (Image Field Directory) within a TIFF file.
}
\keyword{internal}
