% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_getversion}
\alias{cpp_getversion}
\title{Get Current Package Version}
\arguments{
\item{pkg}{s std::string. Default is \code{"IFC"}}
}
\value{
a Rcpp::CharacterVector
}
\description{
Retrieve package version number.
}
\keyword{internal}
