% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_group_df}
\alias{cpp_group_df}
\title{Data Frame Merge Groups with Rcpp}
\arguments{
\item{df}{a DataFrame.}
}
\value{
an IntegerVector with the resulting global grouping factor and a "table" attribute representing the amount of scalar in each resulting level.
}
\description{
Computes global grouping factor from data.frame columns
}
\keyword{internal}
