% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_HSV2RGB}
\alias{cpp_HSV2RGB}
\title{HSV to RGB Conversion}
\source{
see hsv source code from \pkg{grDevices}\cr
Based on HSV_TO_RGB from Foley and Van Dam First Ed. Page 616\cr
See Alvy Ray Smith, Color Gamut Transform Pairs, SIGGRAPH '78
}
\arguments{
\item{*r}{pointer to r}

\item{*g}{pointer to g}

\item{*b}{pointer to b}

\item{h}{double, [0,1], hue. Default is 0.0}

\item{s}{double, [0,1], saturation. Default is 0.0}

\item{v}{double, [0,1], value. Default is 0.0}
}
\description{
Converts a color from HSV [0,1] space to RGB [0,1] space.
}
\keyword{internal}
