% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_int32_to_uint32}
\alias{cpp_int32_to_uint32}
\title{Int32 to Uint32 32bits Conversion}
\arguments{
\item{x}{int32_t.}
}
\description{
Converts 32bits integer from signed to unsigned
}
\keyword{internal}
