% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_int64_to_uint64}
\alias{cpp_int64_to_uint64}
\title{Int64 to Uint64 64bits Conversion}
\arguments{
\item{x}{int64_t.}
}
\description{
Converts 64bits integer from signed to unsigned
}
\keyword{internal}
