% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_inv_compLinLog}
\alias{cpp_inv_compLinLog}
\title{Inverse Compress LinLog Transformation with Rcpp}
\arguments{
\item{x}{NumericVector.}

\item{hyper}{double, value where transition between Lin/Log is applied.}

\item{base}{double, base of Log scale.}

\item{lin_size}{double, size of the linear region as multiple of Log base unit.}
}
\description{
Takes a numeric vector and return its transformation:
- to linear, if abs(x) < lin_size * log(base).
- to exp, if abs(x) > lin_size * log(base).
}
\keyword{internal}
