% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_M_HSV2RGB}
\alias{cpp_M_HSV2RGB}
\title{Matrix HSV to RGB Conversion}
\arguments{
\item{mat}{NumericMatrix, [0,1].}

\item{h}{double, [0,1], hue. Default is 0.0}

\item{s}{double, [0,1], saturation. Default is 0.0}
}
\value{
a NumericVector with 3 dimensions attribute i.e. a 3D array
- 1st Dim is matrix rows count,
- 2nd Dim is matrix cols count,
- 3rd Dim is RGB
}
\description{
Converts grayscale [0,1] mat to 3D rgb array according to hsv space.
hue and saturation determines tint whereas v is given by each element of mat
}
\keyword{internal}
