% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_mark}
\alias{cpp_mark}
\title{Matrix to Matrix Writer According to Mask with Offsets}
\arguments{
\item{A}{a NumericMatrix.}

\item{B}{a NumericMatrix.}

\item{mask}{a NumericMatrix.}

\item{xoff}{x offset in \code{'A'} to start writing \code{'B'}.}

\item{yoff}{x offset in \code{'A'} to start writing \code{'B'}.}

\item{invert}{a logical. Default is \code{false}.
When \code{false}, the default, values of \code{'B'} are written into \code{'A'} when \code{'mask'} is not \code{0.0}.
When \code{true}, values of '\code{1-B}' are written into \code{'A'} when \code{'mask'} is not \code{0.0}.}
}
\description{
Writes matrix \code{'B'} in matrix \code{'A'} according to \code{'mask'}.
}
\keyword{internal}
