% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_mask}
\alias{cpp_mask}
\title{Equal Sized Matrix to Matrix Writer According to Mask}
\arguments{
\item{A}{a NumericMatrix.}

\item{B}{a NumericMatrix.}

\item{mask}{a NumericMatrix.}
}
\description{
Writes matrix \code{'B'} in matrix \code{'A'} according to \code{'mask'}.
If \code{'mask'} is not \code{0.0} \code{'B'} is written, \code{'A'} otherwise.
}
\keyword{internal}
