% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_mpfmatch}
\alias{cpp_mpfmatch}
\title{Multiple Pattern Fixed Matching}
\arguments{
\item{x, pattern}{Nullable Rcpp CharacterVector.}
}
\value{
a bool
}
\description{
String matching with multiple pattern.
}
\details{
equivalent of as.logical(sum(unlist(lapply(pattern, grepl, x = x, fixed = TRUE)))).
}
\keyword{internal}
