% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_NEG_M}
\alias{cpp_NEG_M}
\title{Matrix Neg Logic}
\arguments{
\item{mat}{LogicalMatrix.}
}
\value{
a logical matrix.
}
\description{
This function takes a logical matrix and returns its negation.
}
\keyword{internal}
