% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_offset_to_raw}
\alias{cpp_offset_to_raw}
\title{Offset to Raw Conversion}
\arguments{
\item{x}{double.}

\item{swap}{bool, whether to swap or not.}
}
\description{
Converts offset to raw
}
\keyword{internal}
