% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_raw_to_offset}
\alias{cpp_raw_to_offset}
\title{Offset to Raw Conversion}
\arguments{
\item{x}{RawVector.}

\item{swap}{bool, whether to swap or not.}
}
\description{
Converts raw to offset
}
\keyword{internal}
