% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_readchunk}
\alias{cpp_readchunk}
\title{File Raw Chunk Extraction}
\arguments{
\item{fname}{string, path to file.}

\item{offset}{std::size_t, where to start reading from the beginning.}

\item{nbytes}{uint32_t, number of bytes.}

\item{verbose}{bool, whether to display information (use for debugging purpose). Default is false.}
}
\value{
a RawVector
}
\description{
Reads binary chunk from file.
}
\keyword{internal}
