% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_readFCS}
\alias{cpp_readFCS}
\title{FCS data extraction}
\source{
FCS 3.2 specifications. See, J. Spidlen et al. Data File Standard for Flow Cytometry, Version FCS 3.2. Cytometry A 99 100–102(2021) \doi{10.1002/cyto.a.24225}
}
\arguments{
\item{fname}{string, path to file.}

\item{offset}{std::size_t, offset position of data start}

\item{events}{uint32_t, number of events to read.}

\item{b_typ}{Rcpp::IntegerVector, types of values to read. Allowed are 0 for"A", 1 for "F", 2 for "D", and 3 is "I".}

\item{b_siz}{Rcpp::IntegerVector, number of bytes to extract for each type. Allowed are 0 for 8 1 for 16, 2 for 32 and 3 for 64 bits.\cr
Note that whatever the input is, when 'b_typ' is 1 (float) 'b_siz' will be set to 32 and when 'b_typ' is 2 (double) 'b_siz' will be set to 64.}

\item{b_msk}{Rcpp::IntegerVector, bits to masks when 'b_typ' is 3 (integer). Default is R_NilValue.\cr
When not NULL, it should be of same length as 'b_siz' and contain only [0-64] values.}

\item{swap}{bool, whether to swap bytes or not. Default is false.}
}
\value{
a numeric vector of extracted values.
}
\description{
This function reads data from FCS 3.2 files
}
\keyword{internal}
