% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_replace_non_finite}
\alias{cpp_replace_non_finite}
\title{Non Finite Values Replacement}
\arguments{
\item{V_}{a NumericVector.}

\item{by}{a double used as replcaement value. Default is 0.0}
}
\description{
This function replaces non finite values (NA, NaN -Inf and +Inf)
}
\keyword{internal}
