% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_seqmatch}
\alias{cpp_seqmatch}
\title{Sequence of Strings Matching}
\arguments{
\item{x, y}{StringVector to match}
}
\value{
the index (starting at 1) when a match has been found. Otherwise 0.
}
\description{
Match a sequence of strings in another string
}
\details{
smallest sequence will be searched into the largest one.
}
\keyword{internal}
