% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_uint32_to_int32}
\alias{cpp_uint32_to_int32}
\title{Uint32 to Int32 32bits Conversion}
\arguments{
\item{x}{uint32_t.}
}
\description{
Converts 32bits integer from unsigned to signed
}
\keyword{internal}
