% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_uint64_to_int64}
\alias{cpp_uint64_to_int64}
\title{Uint64 to Int64 64bits Conversion}
\arguments{
\item{x}{uint64_t.}
}
\description{
Converts 64bits integer from unsigned to signed
}
\keyword{internal}
