% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_upscale}
\alias{cpp_upscale}
\title{Image Upscaling}
\arguments{
\item{mat}{Matrix (Raw, Integer Logical or Numeric).}

\item{scale}{double, giving the scaling factor. Default is 1.0 for no change.}

\item{resize}{bool, whether to resize upscaled mat to its original dimension. Default is false.}

\item{interpolation}{uint8_t, type of interpolation applied on output (2 = "bilinear", 3 = "bicubic" a = -0.5). Default is 2. Values other than 2 or 3 will result in nearest neighbor interpolation.}

\item{add_noise}{logical, if true adds normal noise when at least one new dimension is larger than original mat dimensions. Rcpp::rnorm() function is used. Default is true.}

\item{bg}{double, mean value of the background added if add_noise is true. Default is 0.0.}

\item{sd}{double, standard deviation of the background added if add_noise is true. Default is 0.0.}
}
\value{
a rescaled matrix.
}
\description{
Function that upscales mat according to scale.
}
\keyword{internal}
