% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_v_int32_to_uint32}
\alias{cpp_v_int32_to_uint32}
\title{Vectorize Int32 to Uint32 32bits Conversion}
\arguments{
\item{V}{a NumericVector}
}
\description{
Converts 32bits vector of integers from unsigned to signed
}
\keyword{internal}
