% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_v_int64_to_uint64}
\alias{cpp_v_int64_to_uint64}
\title{Vectorize Int64 to Uint64 64bits Conversion}
\arguments{
\item{V}{a NumericVector}
}
\description{
Converts 64bits vector of integers from unsigned to signed
}
\keyword{internal}
