% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_writeBMP}
\alias{cpp_writeBMP}
\title{BMP Writer}
\arguments{
\item{image, }{a [0,1] normalized image matrix or 3D array. If 3D array, 3rd dimension should be of length 1 or 3.}
}
\description{
Transforms 3D [0,1] image to uncompressed bmp
}
\keyword{internal}
