% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_writeIFD}
\alias{cpp_writeIFD}
\title{IFD Tag Writer}
\arguments{
\item{img}{RawVector, an encoded image. It should contain 'dims', 'what' and 'comp' attributes.}

\item{tags}{List, extra tags to be included to IFD. Expecting a list whose sub-elements are list containing:\cr
-'tag' uint16_t, tag number, it should be [1-65535],\cr
-'typ' uint16_t typ number, it should be [1-12],\cr
-'map' SEXP vector of values to write, it should not be empty and should be even for 'typ' 5 and 10.}

\item{offset}{uint32_t, position of the IFD beginning. Default is \code{0}.}

\item{endianness}{std::string, "little" or "big".}

\item{rgb}{bool, whether to write channels as rgb. Default is \code{false}.}

\item{last}{bool, whether IFD is the last one. Default is \code{false}.}

\item{verbose}{bool, whether to display information (use for debugging purpose). Default is \code{false}.}
}
\value{
a RawVector
}
\description{
Writes TIFF IFD (Image Field Directory).
}
