% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportToReport.R
\name{CreateGraphReport}
\alias{CreateGraphReport}
\title{Graph Report Generation}
\usage{
CreateGraphReport(
  obj,
  selection,
  onepage = TRUE,
  color_mode = c("white", "black")[1],
  add_key = "panel",
  precision = c("light", "full")[1],
  trunc_labels = 38,
  trans = "asinh",
  bin,
  viewport = "ideas",
  backend = "lattice",
  display_progress = TRUE,
  ...
)
}
\arguments{
\item{obj}{an `IFC_data` object extracted with features extracted.}

\item{selection}{indices of desired graphs. It can be provided as an integer vector or as a matrix.\cr
In such case, the layout of the matrix will reflect the layout of the extracted graphs.\cr
NA value will result in an empty place.\cr
Otherwise, when 'selection' is provided as a vector not identical to seq_along(obj$graphs), 'onepage' parameter will be set to FALSE.\cr
Note that indices are read from left to right, from top to bottom. Default is missing for extracting all graphs.}

\item{onepage}{whether to generate a pdf with all graphs on one page or not. Default is TRUE.}

\item{color_mode}{Whether to extract colors from obj in white or black mode. Default is "white".}

\item{add_key}{whether to draw a "global" key under title or in the first "panel" or "both". Default is "panel".\cr
Accepted values are either: FALSE, "panel", "global", "both" or c("panel", "global").\cr
Note that it only applies when display is seen as overlaying populations.}

\item{precision}{when graphs is a 2D scatter with population overlay, this argument controls amount of information displayed. Default is "light".\cr
-"light", the default, will only display points of same coordinates that are among the other layers.\cr
-"full" will display all the layers.}

\item{trunc_labels}{maximum number of characters to display for labels. Default is 38.}

\item{trans}{name of the transformation function for density graphs. If missing the default, the BasePop[[1]]$densitytrans, if any, will be retrieved, otherwise "asinh" will be used.}

\item{bin}{default number of bin used for histogram. Default is missing.}

\item{viewport}{Either "ideas", "data" or "max" defining limits used for the graph. Default is "ideas".\cr
-"ideas" will use same limits as the one defined in ideas.\cr
-"data" will use data to define limits.\cr
-"max" will use data and regions drawn to define limits.\cr
Alternatively, "x-" or "y-" can be appended to control limits adjustment for each axis separately.}

\item{backend}{backend used for drawing. Allowed are "lattice", "base", "raster". Default is "lattice".\cr
-"lattice" is the original one used in \pkg{IFC} using \pkg{lattice},\cr
-"base" will produce the plot using \pkg{base},\cr
-"raster" uses "base" for plotting but 2D graphs points will be produced as \code{\link[graphics]{rasterImage}}.\cr
This has the main advantage of being super fast allowing for plotting a huge amount of points while generating smaller objects (in bytes).
However, plot quality is impacted with "raster" method and resizing can lead to unpleasant looking.}

\item{display_progress}{whether to display a progress bar. Default is TRUE.}

\item{...}{other parameters to be passed.}
}
\value{
a list with onepage, layout, layout_matrix, graphs, grobs, and stats.
}
\description{
Generates graph report (plot + statistics) from `IFC_data` object.
}
\keyword{internal}
