% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ome.R
\name{createOME}
\alias{createOME}
\title{OME XML Writer}
\usage{
createOME(
  dims,
  param,
  name,
  date,
  what = "int16",
  endianness = .Platform$endian
)
}
\arguments{
\item{dims}{image dimensions [h,w,c,f]\cr
with h=height, w=width, c=channel, f=frame.}

\item{param}{object of class `IFC_param`, containing extraction parameters defined by \code{\link{objectParam}}.}

\item{name}{name of the image. When missing, the default it will not be incorporated.}

\item{date}{date of the image. Should be formatted as \verb{"\%Y:\%m:\%d \%H:\%M:\%S"}, see \code{\link[base]{format.Date}}. When missing, the default it will not be incorporated.}

\item{what}{bits mode used to store image. Default is \code{"int16"}. Allowed are \code{"uint8"}, \code{"int8"}, \code{"uint16"}, \code{"int16"}, \code{"uint32"}, \code{"int32"}, \code{"float"} and \code{"double"}.}

\item{endianness}{The endian-ness ("big" or "little") of the return object. Default is .Platform$endian.}
}
\value{
An OME xml.
}
\description{
Creates OME xml
}
\keyword{internal}
