% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportToGallery.R
\name{CreateSheet}
\alias{CreateSheet}
\title{Sheet Creation}
\usage{
CreateSheet(
  imgs,
  layout,
  main = character(0),
  grid = c(c(length(imgs), 1)[c(!byrow, byrow)], c(length(imgs), 1)[c(byrow, !byrow)]),
  byrow = TRUE,
  add_channels = TRUE,
  add_lines = c(2, 0),
  add_ids = 0,
  order_ids = FALSE,
  scale = list(),
  fv = NULL,
  fv_fmt = "\%05.3f",
  fv_color = "darkblue",
  id_color = "yellow",
  ln_color = "grey20",
  bn_color = "grey6",
  rz_color = NA_character_
)
}
\arguments{
\item{imgs}{`IFC_img` / `IFC_msk` list as extracted by \code{\link{ExtractImages_toMatrix}} or \code{\link{ExtractMasks_toMatrix}} functions with \code{'mode'="rgb"}.}

\item{layout}{a character vector of [acquired channels + 'composite' images] members to export. Default is missing to export everything.
Note that members can be missing to be removed from final montage.
Note that members not found will be automatically removed and a warning will be thrown.}

\item{main}{main title that will be displayed on top center of the image.
If too large it will be clipped. Default is \code{character(0)}.}

\item{grid}{grid size of the image montage. Default is \code{c(c(length(imgs),1)[c(byrow,!byrow)],c(length(imgs),1)[c(!byrow,byrow)])}.}

\item{byrow}{whether montage is filled row-wise or not. Default is \code{TRUE}.}

\item{add_channels}{whether to add channels names. Default is \code{TRUE}.}

\item{add_lines}{a length=2 finite integer vector, size of separating lines between objects. Default is \code{c(2,0)}, 1st value control horizontal space and 2nd vertical space.
If add_lines < 1, no separating lines are added.}

\item{add_ids}{integer, index of column to mark objects ids number. Default is \code{0}.
If add_ids < 1, no ids are added. It will also be the one use for \code{'fv'}.}

\item{order_ids}{whether to sort images. Default is \code{FALSE}.}

\item{scale}{a named list to pass to \code{addScaleBar} function. Default is \code{list()}.
\code{'scale'=TRUE}, can eventually be used to draw scale with default parameters.}

\item{fv}{feature values to display. Default is \code{NULL}. It should be a named vector whose named are objects' ids.}

\item{fv_fmt}{\code{'fmt'} argument to \code{\link{sprintf}} for \code{'fv'} formatting. Default is \code{"\%05.3f"}.}

\item{fv_color}{color for \code{'fv'}. Default is \code{"darkblue"}.}

\item{id_color}{color for ids. Default is \code{"yellow"}.}

\item{ln_color}{background color for separating lines. Default is \code{"grey20"}.}

\item{bn_color}{background color for main and channels. Default is \code{"grey6"}. Foreground color will be \code{"white"} or \code{"black"} depending on \code{'bn_color'} luminance.}

\item{rz_color}{color used when resizing is needed fo fill empty space in the montage. Default is \code{NA_character_} to use image color.}
}
\value{
a rgb array
}
\description{
Creates contact sheet of `IFC_img` / `IFC_msk` objects
}
\keyword{internal}
