% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death_universities.R
\name{death_universities}
\alias{death_universities}
\title{Returns list of universities which underwent a fraternity death in 2014-2019.}
\usage{
death_universities()
}
\description{
\code{death_universities} returns a list of universities in which a moratorium has occurred due to death in addition to never-treated (but had death) schools.
}
\details{
This function simply returns a list of universities. Does not rely on anything.
}
\examples{
death_universities()

}
