% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death_untreated_universities.R
\name{death_untreated_universities}
\alias{death_untreated_universities}
\title{Returns list of universities which underwent a fraternity death in 2014-2019 but no moratorium.}
\usage{
death_untreated_universities()
}
\description{
\code{death_untreated_universities} returns a list of universities in which a moratorium has occurred due to death in addition to never-treated schools.
}
\details{
This function simply returns a list of schools.
}
\examples{
death_universities()

}
