% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{densCols}
\alias{densCols}
\title{Colors for Smooth Density Plots}
\source{
derived from \pkg{grDevices} R Core Team, Florian Hahne at FHCRC, originally
}
\usage{
densCols(
  x,
  y = NULL,
  xlim = range(x, na.rm = TRUE, finite = TRUE),
  ylim = range(y, na.rm = TRUE, finite = TRUE),
  nbin = 128,
  colramp = colorRampPalette(c("blue", "green", "red")),
  transformation = "asinh"
)
}
\description{
Helper to map density to colors
}
\keyword{internal}
