% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{dev_close}
\alias{dev_close}
\title{Graphical Devices Shutdown}
\usage{
dev_close(dv = dev.list())
}
\arguments{
\item{dv}{a graphical device list. Default is dev.list()}
}
\value{
invisible(NULL)
}
\description{
Closes all open devices
}
\keyword{internal}
