% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportToReport.R
\name{DisplayReport}
\alias{DisplayReport}
\title{Graphical and Statistic Report Display}
\usage{
DisplayReport(obj, display_progress = TRUE, ...)
}
\arguments{
\item{obj}{an `IFC_data` object extracted with features extracted.}

\item{display_progress}{whether to display a progress bar. Default is TRUE.}

\item{...}{other parameters to be passed.}
}
\value{
It invisibly returns NULL
}
\description{
Displays report from `IFC_data` object.
}
\keyword{internal}
