% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractToFuns.R
\name{dotsParam}
\alias{dotsParam}
\title{Shortcut for Parameters Retrieval}
\usage{
dotsParam(..., export = "matrix", mode = "raw")
}
\arguments{
\item{\dots}{arguments to be passed to \code{\link{objectExtract}} for `IFC_param` retrieval.\cr
\strong{/!\\} If not any of \code{'fileName'}, \code{'info'} and \code{'param'} can be found in \code{'...'} then \code{attr(offsets, "fileName_image")} will be used as \code{'fileName'} input parameter to pass to \code{\link{objectParam}}.}

\item{export}{(\code{\link{objectParam}} argument) format mode export. Either \code{"file"}, \code{"matrix"}, \code{"base64"}. Default is \code{"matrix"}.}

\item{mode}{(\code{\link{objectParam}} argument) color mode export. Either \code{"raw"}, \code{"rgb"}, \code{"gray"}. Default is \code{"rgb"}.}
}
\value{
an object of class `IFC_param`.
}
\description{
Function to retrieve parameters from input.
}
\details{
If \code{'param'} is provided in \code{'...'}, \code{'param$export'<-'export'} and \code{'param$mode'<-'mode'} \strong{only} will be overwritten.
}
\note{
Arguments of \code{\link{objectParam}} will be deduced from input arguments.
}
\keyword{internal}
