% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_utils.R
\name{dual_coords}
\alias{dual_coords}
\title{Dual Coordinates}
\usage{
dual_coords(coords, part, xlim, ylim)
}
\arguments{
\item{coords}{a list containing x and y numeric vectors of at least 2 elements each, if not 1st element will be repeated twice.}

\item{part}{a string referencing sync part, as returned by \code{sync_part()}.}

\item{xlim}{x-limits. Default is missing to use \code{sync_lims()}, otherwise it will be \code{sync_lims(xlim)}.}

\item{ylim}{y-limits. Default is missing to use \code{sync_lims()}, otherwise it will be \code{sync_lims(ylim)}.}
}
\value{
coordinates for the corresponding \code{part}
}
\description{
Helper to construct coordinates for "dual" sync.
}
\details{
1st x should be the coordinate where parts 1 and 2 are separated.
1st y should be the y-coordinate for part 1 and 2nd y the one for part 2.
Then min/max xlim will be distributed to each part:\cr
-1: x = x[1], min(xlim); y = y[1], y[1]\cr
-2: x = x[1], max(xlim); y = y[2], y[2]\cr
otherwise, x = x[1], x[2]; y = y[1], y[2].\cr
In addition, \code{-Inf} x values will be replaced by \code{min(xlim)} and \code{+Inf} x values by \code{max(xlim)}, same for y values and \code{'ylim'}.
}
\keyword{internal}
