% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writetiff.R
\name{editMetadata}
\alias{editMetadata}
\title{Metadata Editor}
\usage{
editMetadata(fileName, metadata = NULL)
}
\arguments{
\item{fileName}{path to file.}

\item{metadata}{string used to fill tag "270" of first IFD of 'fileName'. Default is \code{NULL}.}
}
\value{
It invisibly returns path of edited file.
}
\description{
Edits metadata in TIFF file
}
\details{
/!\ file will be modified /!\
It will only work for 1st time edition of TIFF files produced by \pkg{IFC}.
}
\keyword{internal}
