% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writetiff.R
\name{editTag}
\alias{editTag}
\title{Image Field Directory Tag Editor}
\usage{
editTag(IFD, which, tag, value, force = FALSE)
}
\arguments{
\item{IFD}{an object of class `IFC_ifd_list` extracted by \code{\link{getIFD}}.}

\item{which}{scalar, integer (index) or the name of 'IFD' sub-element to edit 'tag' to.}

\item{tag}{scalar, integer (index) or the name of the desired 'tag' in IFD[[which]].}

\item{value}{value used to fill the 'tag'. For the moment only string is supported.}

\item{force}{whether to force edition. Needed when 'tag' value already exceed 4 bytes. Default is \code{FALSE}.}
}
\value{
It invisibly returns path of edited file.
}
\description{
Edits tag value in file IFDs (Image Field Directory).
}
\details{
/!\ file will be modified /!\
/!\ When \code{'force'} needs to be \code{TRUE}, 'value' will be appended to file \strong{but} former tag 'value' will still be present in file \strong{and not} erased.\cr
Will only work when 'tag' is present in TIFF file.\cr
Will only work for TIFF files produced by \pkg{IFC}
}
\keyword{internal}
