% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_utils.R
\name{endPB}
\alias{endPB}
\title{Progress Bar Terminator}
\usage{
endPB(pb)
}
\arguments{
\item{pb}{an object of class "txtProgressBar" "winProgressBar" or "Progress".}
}
\value{
an object of class "txtProgressBar" "winProgressBar" or "Progress".
}
\description{
Closes a progress bar.
}
\keyword{internal}
