% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{escape_entities}
\alias{escape_entities}
\title{XML Entities Protection}
\usage{
escape_entities(text)
}
\arguments{
\item{text}{value return by xml2::as_list. Default is 5.}
}
\value{
a character vector where xml entities have been escaped.
}
\description{
Helper to escape xml entities.
}
\details{
entities will be replaced by:\cr
-& to "&amp;"\cr
-> to "&gt;"\cr
-< to "&lt;"\cr
-' to "&apos;"\cr
-" to "&quot;"\cr
}
\keyword{internal}
