% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_study_graph.R
\name{event_study_graph}
\alias{event_study_graph}
\title{Graphs an event study with no labs.}
\usage{
event_study_graph(x, window_size, reference_period = window_size)
}
\arguments{
\item{x}{a regression fixest object.}

\item{window_size}{the length of the event window.}

\item{reference_period}{the period to be omitted, defaults to period before. This is found by counting from last period to period wanted.}
}
\description{
\code{event_study_graph} returns a ggplot object of the event study graph.
}
\details{
This function relies on dplyr, ggplot, and tibble packages to work.
}
\examples{
event_study_graph(x, 8, 8)

}
