% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_study_stagger.R
\name{event_study_stagger}
\alias{event_study_stagger}
\title{Creates the Event Study tibble for analysis at Daily Level Data.}
\usage{
event_study_stagger(data, leads_lags, clusters)
}
\arguments{
\item{data}{an object/dataframe/tibble (daily_crime).}

\item{leads_lags}{how big the event window is.}

\item{clusters}{a number for how many days you want clustered together.}
}
\description{
\code{event_study_stagger} returns a tibble ready for event study analysis in units of days with staggered adoption
}
\details{
This function creates the event study tibble needed to do analysis in event study framework.
}
\examples{
event_study_stagger(daily_crime, leads_lags = 8, bin_length =7)

}
