% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{expand_list}
\alias{expand_list}
\title{XML Node List Expansion}
\usage{
expand_list(x, max = 5)
}
\arguments{
\item{x}{value return by xml2::as_list. Default is 5.}

\item{max}{maximum number of recurrence into subnodes.}
}
\description{
Helper to stringify XML node.
}
\keyword{internal}
