% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportToBATCH.R
\name{ExportToBATCH}
\alias{ExportToBATCH}
\title{Batch File Writer}
\usage{
ExportToBATCH(batch)
}
\arguments{
\item{batch}{list of batch nodes as created by \code{\link{buildBatch}}.}
}
\value{
It invisibly returns full path of xml batch file.
}
\description{
Writes an XML file to batch files
}
