% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractToFuns.R
\name{ExtractMasks_toBase64}
\alias{ExtractMasks_toBase64}
\title{Shortcut for Batch Masks Extraction to Base64}
\usage{
ExtractMasks_toBase64(
  ...,
  objects,
  offsets,
  display_progress = TRUE,
  mode = c("rgb", "gray")[1]
)
}
\arguments{
\item{...}{arguments to be passed to \code{\link{objectExtract}} with the exception of \code{ifd} and \code{bypass}(=\strong{TRUE}).\cr
\strong{/!\\} If not any of '\code{fileName}', '\code{info}' and '\code{param}' can be found in \code{\dots} then \code{attr(offsets, "fileName_image")} will be used as '\code{fileName}' input parameter to pass to \code{\link{objectParam}}.}

\item{objects}{integer vector, IDEAS objects ids numbers to use.
This argument is not mandatory, if missing, the default, all objects will be used.}

\item{offsets}{object of class `IFC_offset`.
This argument is not mandatory but it may allow to save time for repeated image export on same file.\cr
If '\code{offsets}' are not provided extra arguments can also be passed with \code{\dots} to \code{\link{getOffsets}}.}

\item{display_progress}{whether to display a progress bar. Default is TRUE.}

\item{mode}{(\code{\link{objectParam}} argument) color mode export. Either "rgb", "gray". Default is "rgb".}
}
\value{
A list of base64 encoded images corresponding to objects extracted.
}
\description{
Function to shortcut extraction, normalization and eventually colorization of masks to base64 ! excludes image.
}
\details{
arguments of \code{\link{objectExtract}} will be deduced from \code{\link{ExtractMasks_toBase64}} input arguments.\cr
If '\code{param}' is provided 'in \code{\dots}, \code{param$export}(=\strong{"base64"}) and \code{param$mode}(=\code{mode}) \strong{only} will be overwritten.
}
