% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{extractStats}
\alias{extractStats}
\title{Statistics Extraction}
\usage{
extractStats(obj, feat_name, trans = "P")
}
\arguments{
\item{obj}{an `IFC_data` object}

\item{feat_name}{a feature name.}

\item{trans}{character string describing transformation used and its parameters. See parseTrans().}
}
\value{
a data.frame with computed statistics.
}
\description{
Extracts statistics from `IFC_data` object
}
\keyword{internal}
