% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastCbind.R
\name{fastCbind}
\alias{fastCbind}
\title{Combine by Columns}
\usage{
fastCbind(obj1, obj2, add_id = FALSE)
}
\arguments{
\item{obj1}{an object either a data.frame, a list or something coercible to numeric matrix.}

\item{obj2}{an object either a data.frame, a list or something coercible to numeric matrix.}

\item{add_id}{a bool determining if 1st column of returned object should be given 1 to nrow integers}
}
\value{
a combined object
}
\description{
Helper to combine by columns
}
\details{
if obj1 or obj2 is a data.frame returned object will inherit class of obj1 or obj2 respectively in this order.
/!\ if obj1 or obj2 needs to be coerced take care that it can not handle factor during object coercion
}
\keyword{internal}
