% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs.R
\name{FCS_check_keywords}
\alias{FCS_check_keywords}
\title{FCS Keyword Checker}
\usage{
FCS_check_keywords(text, delimiter = raw(1), version = 3, fun = "warning", ...)
}
\arguments{
\item{text}{a named list of keywords values.}

\item{delimiter}{byte used as delimiter. Default is raw(1).}

\item{version}{version to check keywords compliance against. Default is 3.0,}

\item{fun}{function to execute when mandatory parameters are not met. Default is "warning". Allowed are "stop","warning","message","return".}

\item{...}{other arguments to be passed.}
}
\description{
Helper to check that FCS keyword/value pairs are compliant with specification
}
\keyword{internal}
