% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs.R
\name{FCS_merge_dataset}
\alias{FCS_merge_dataset}
\title{FCS Object Data Sets Merging}
\usage{
FCS_merge_dataset(fcs, ...)
}
\arguments{
\item{fcs}{`IFC_fcs` object as extracted by readFCS().}

\item{...}{other arguments to be passed.}
}
\value{
a list of list containing:\cr
- header, list of header information corresponding to 'options'\cr
- delimiter, unique character used to separate keyword - values\cr
- text, list of keywords values,\cr
- data, data.frame of values.
}
\description{
Merges FCS data object with various data sets.
}
\details{
in data can contain extra columns named 'import_file' and 'import_subfile' intended to allow file/dataset identification
}
\keyword{internal}
