% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs.R
\name{FCS_to_data}
\alias{FCS_to_data}
\title{FCS Object Converter}
\usage{
FCS_to_data(fcs, ...)
}
\arguments{
\item{fcs}{`IFC_fcs` object as extracted by readFCS().}

\item{...}{other arguments to be passed.}
}
\value{
A named list of class `IFC_data`, whose members are:\cr
-description, a list of descriptive information,\cr
-Merged_fcs, character vector of path of files used to create fcs, if input was a merged,\cr
-fileName, path of fileName input,\cr
-fileName_image, path of .cif image fileName is referring to,\cr
-features, a data.frame of features,\cr
-features_def, a describing how features are defined,\cr
-graphs, a list of graphical elements found,\cr
-pops, a list describing populations found,\cr
-regions, a list describing how regions are defined,\cr
-images, a data.frame describing information about images,\cr
-offsets, an integer vector of images and masks IFDs offsets,\cr
-stats, a data.frame describing populations count and percentage to parent and total population,\cr
-checksum, a checksum integer.
}
\description{
Converts FCS data object to `IFC_data` object.
}
\details{
in data can contain extra columns named 'import_file' and 'import_subfile' intended to allow file/dataset identification
}
\keyword{internal}
