% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redefine_utils.R
\name{feature_namer}
\alias{feature_namer}
\title{Feature Default Name Computation}
\usage{
feature_namer(feat_def)
}
\arguments{
\item{feat_def}{list containing feature definition}
}
\value{
a string with default name
}
\description{
Helper to compute default name of a feature.
}
\keyword{internal}
