% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{fetch_seed}
\alias{fetch_seed}
\title{Seed Instructions Extraction}
\usage{
fetch_seed(seed = NA_integer_)
}
\arguments{
\item{seed}{a list of elements to pass to \link[base]{set.seed} or a single value, interpreted as an integer, or NULL. NA_integer_ or list(seed = NA_integer_) can be used to prevent 'seed' argument from being passed to \link[base]{set.seed}. Default is NA_integer_.}
}
\description{
Retrieve seed instructions from input
}
\keyword{internal}
