% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_utils.R
\name{formatdate}
\alias{formatdate}
\title{Date Converter}
\usage{
formatdate(
  x,
  iso = FALSE,
  tryFormats = c("\%m/\%d/\%Y \%I:\%M:\%S \%p", "\%d/\%m/\%Y \%H:\%M:\%S",
    "\%Y:\%m:\%d \%H:\%M:\%S", "\%Y-\%m-\%dT\%H:\%M:\%SZ"),
  tz = "UTC"
)
}
\arguments{
\item{x}{a string.}

\item{iso}{whether returned value should be ISO8601. Default is \code{FALSE}.}

\item{tryFormats}{format to try. See \code{\link[base]{format.Date}}. Default is \code{c("\%m/\%d/\%Y \%I:\%M:\%S \%p", "\%d/\%m/\%Y \%H:\%M:\%S", "\%Y:\%m:\%d \%H:\%M:\%S", "\%Y-\%m-\%dT\%H:\%M:\%SZ")}.}

\item{tz}{time zone name. See \code{\link[base]{format.Date}}. Default is "UTC".}
}
\value{
formatted date
}
\description{
Formats date.
}
\keyword{internal}
