% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingML.R
\name{fromXML2_gating}
\alias{fromXML2_gating}
\title{GatingML Conversion from XML2}
\usage{
fromXML2_gating(xml_nodes, type = "rect")
}
\arguments{
\item{xml_nodes}{a set of xml_nodes}

\item{type}{type of the gate to extract}
}
\value{
a list of:\cr
-region, region information\cr
-pop, population information
}
\description{
Helper to convert boolean and graphical pops and corresponding regions from XML nodes in GatingML files.
}
\keyword{internal}
