% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromXML2_pops.R
\name{fromXML2_pops}
\alias{fromXML2_pops}
\title{XML Populations Parser}
\usage{
fromXML2_pops(
  xml_doc,
  caller = character(0),
  title_progress = "",
  display_progress = FALSE
)
}
\arguments{
\item{xml_doc}{an `xml_document` object.}

\item{caller}{a string, name of the calling function. Default is \code{character(0)}.}

\item{title_progress}{title displayed in progress bar. Default is \code{""}.}

\item{display_progress}{whether to display a progress bar. Default is \code{FALSE}.}
}
\value{
a list of populations.
}
\description{
Helper to parse populations from XML document.
}
\keyword{internal}
