% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{get_coordmap_raw}
\alias{get_coordmap_raw}
\title{Device Raw Coordinates}
\source{
computes drawing region in a similar way as shiny:::getPrevPlotCoordmap()
}
\usage{
get_coordmap_raw()
}
\description{
Helper to extract current device plotting region
}
\keyword{internal}
