% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsStyle.R
\name{get_pop_style}
\alias{get_pop_style}
\title{Population Style}
\usage{
get_pop_style(x)
}
\arguments{
\item{x}{type of population.}
}
\value{
a named UTF-8 character
}
\description{
Gets population style
}
\keyword{internal}
