% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_pops_stats}
\alias{get_pops_stats}
\title{Statistics Extraction from Populations}
\usage{
get_pops_stats(pops, objcount = 0)
}
\arguments{
\item{pops}{list of populations}

\item{objcount}{total number of objects.}
}
\value{
a data.frame with computed statistics.
}
\description{
Extracts populations statistics
}
\keyword{internal}
