% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{get_ylim}
\alias{get_ylim}
\title{Histogram y-Axis Limits Constructor}
\source{
derived from \pkg{lattice} from Deepayan Sarkar
}
\usage{
get_ylim(x, type, br, include.lowest = TRUE, right = TRUE)
}
\description{
Helper to extract ylim
}
\keyword{internal}
