% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getASSIST.R
\name{getASSIST}
\alias{getASSIST}
\title{ASSIST Database Extraction}
\usage{
getASSIST(fileName, ...)
}
\arguments{
\item{fileName}{path to file..}

\item{...}{other arguments to be passed.}
}
\value{
a list of class `IFC_assist` of parsed ASSIST tests database.
}
\description{
Retrieves ASSIST tests values stored within .cif / .rif files.
}
\keyword{internal}
