% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeaturesValues.R
\name{getFeaturesValues}
\alias{getFeaturesValues}
\title{Features Values Extraction}
\usage{
getFeaturesValues(
  features,
  features_def,
  operators = c("+", "-", "*", "/", "(", ")", "ABS", "COS", "SIN", "SQR", "SQRT"),
  split = "|",
  ...
)
}
\arguments{
\item{features}{a data.frame of features, typically an object of class `IFC_features`.}

\item{features_def}{a list of features definitions, typically an object of class `IFC_features_def`.}

\item{operators}{operators used. Default is c("+", "-", "*", "/", "(", ")", "ABS", "COS", "SIN", "SQR", "SQRT").}

\item{split}{string used for splitting. Default is "|".}
}
\value{
a data.frame of features values.
}
\description{
Extracts features values according to features definitions
}
\keyword{internal}
