% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFileExt.R
\name{getFileExt}
\alias{getFileExt}
\title{File Extension Retrieval}
\source{
derived from file_ext() in \pkg{tools}, R Core Team, Kurt Hornik and Friedrich Leisch.
}
\usage{
getFileExt(x)
}
\arguments{
\item{x}{path to file.}
}
\value{
the file extension in lower case if found. Otherwise, "".
}
\description{
Determines extension from alpha numeric file name.
}
\keyword{internal}
