% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLuminance.R
\name{getLuminance}
\alias{getLuminance}
\title{Luminance Computation}
\usage{
getLuminance(color)
}
\arguments{
\item{color}{a character string.}
}
\description{
Determines color's luminance.
}
\keyword{internal}
