% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_mapping.R
\name{gof_mapping}
\alias{gof_mapping}
\title{Returns tibble of the goodness of fit mappings for tables.}
\usage{
gof_mapping()
}
\description{
\code{gof_mapping} returns a tibble of goodness of fit mappings for the tables.
}
\details{
This function simply returns a tibble. Does not rely on anything besides having the tibble package installed.
}
\examples{
gof_mapping()

}
