% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{hist_constr}
\alias{hist_constr}
\title{Histogram Constructor}
\source{
derived from \pkg{lattice} from Deepayan Sarkar
}
\usage{
hist_constr(x, breaks, include.lowest = TRUE, right = TRUE,
  plot = FALSE)
}
\description{
Helper to construct histogram
}
\keyword{internal}
