% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifc-package.R
\docType{package}
\name{ifc-package}
\alias{ifc}
\alias{ifc-package}
\title{ifc: International Fixed Calendar Date Class}
\description{
The \code{ifc} package implements the International Fixed Calendar as a proper
S3 vector class built on the \href{https://vctrs.r-lib.org/}{vctrs} framework. The IFC divides each year
into 13 months of exactly 28 days, with a Year Day appended at year's end
(and a Leap Day after IFC June 28 in leap years). Because each month always
starts on Sunday and ends on Saturday, the day-of-week for any IFC date
is fully deterministic from the day-of-month number.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://k5cents.github.io/ifc/}
  \item \url{https://github.com/k5cents/ifc}
  \item Report bugs at \url{https://github.com/k5cents/ifc/issues}
}

}
\author{
\strong{Maintainer}: Kiernan Nicholls \email{k5cents@gmail.com} (\href{https://orcid.org/0000-0002-9229-7897}{ORCID}) [copyright holder]

}
\keyword{internal}
