% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{ifc_coerce}
\alias{ifc_coerce}
\alias{as.Date.ifc_date}
\alias{as.POSIXct.ifc_date}
\alias{as.POSIXlt.ifc_date}
\alias{as.integer.ifc_date}
\alias{as.double.ifc_date}
\alias{as.character.ifc_date}
\title{Coerce an IFC date to a base R date type}
\usage{
\method{as.Date}{ifc_date}(x, ...)

\method{as.POSIXct}{ifc_date}(x, tz = "UTC", ...)

\method{as.POSIXlt}{ifc_date}(x, tz = "UTC", ...)

\method{as.integer}{ifc_date}(x, ...)

\method{as.double}{ifc_date}(x, ...)

\method{as.character}{ifc_date}(x, ...)
}
\arguments{
\item{x}{An \code{ifc_date} vector.}

\item{...}{Ignored.}

\item{tz}{Time zone string passed to \code{as.POSIXct} / \code{as.POSIXlt}.}
}
\value{
The corresponding base R type.
}
\description{
Coerce an IFC date to a base R date type
}
