% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{ifc_datetime_accessors}
\alias{ifc_datetime_accessors}
\alias{ifc_hour}
\alias{ifc_minute}
\alias{ifc_second}
\title{Extract time components from an IFC datetime}
\usage{
ifc_hour(x)

ifc_minute(x)

ifc_second(x)
}
\arguments{
\item{x}{An \code{ifc_datetime} vector.}
}
\value{
\code{ifc_hour()}: integer 0--23.

\code{ifc_minute()}: integer 0--59.

\code{ifc_second()}: double 0--60 (60 for leap seconds).
}
\description{
Extract time components from an IFC datetime
}
\examples{
x <- ifc_datetime("2024-07-14 09:30:45", tz = "UTC")
ifc_hour(x)
ifc_minute(x)
ifc_second(x)
}
