% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{ifc_datetime_coerce}
\alias{ifc_datetime_coerce}
\alias{as.POSIXct.ifc_datetime}
\alias{as.POSIXlt.ifc_datetime}
\alias{as.Date.ifc_datetime}
\alias{as.character.ifc_datetime}
\alias{as.double.ifc_datetime}
\title{Coerce an IFC datetime to a base R datetime type}
\usage{
\method{as.POSIXct}{ifc_datetime}(x, tz = NULL, ...)

\method{as.POSIXlt}{ifc_datetime}(x, tz = NULL, ...)

\method{as.Date}{ifc_datetime}(x, tz = NULL, ...)

\method{as.character}{ifc_datetime}(x, ...)

\method{as.double}{ifc_datetime}(x, ...)
}
\arguments{
\item{x}{An \code{ifc_datetime} vector.}

\item{tz}{Time zone string. If \code{NULL} (default), uses the timezone stored
in \code{x}.}

\item{...}{Ignored.}
}
\value{
The corresponding base R type.
}
\description{
Coerce an IFC datetime to a base R datetime type
}
