% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\docType{data}
\name{IFC_MONTH_ABBR}
\alias{IFC_MONTH_ABBR}
\title{IFC month abbreviations (1–13)}
\format{
An object of class \code{character} of length 13.
}
\usage{
IFC_MONTH_ABBR
}
\description{
IFC month abbreviations (1–13)
}
\keyword{datasets}
