% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{ifc_now}
\alias{ifc_now}
\title{Current date-time in the IFC}
\usage{
ifc_now(tz = "UTC")
}
\arguments{
\item{tz}{Time zone string. Defaults to \code{"UTC"}.}
}
\value{
An \code{ifc_datetime} of length 1 representing right now.
}
\description{
A convenience wrapper around \code{ifc_datetime(Sys.time(), tz = tz)}.
}
\examples{
ifc_now()
ifc_now("America/New_York")
}
