% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{ifc_today}
\alias{ifc_today}
\title{Today's date in the IFC}
\usage{
ifc_today()
}
\value{
An \code{ifc_date} of length 1 representing today.
}
\description{
A convenience wrapper around \code{ifc_date(Sys.Date())}.
}
\examples{
ifc_today()
}
