% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\docType{data}
\name{IFC_WDAY_ABBR}
\alias{IFC_WDAY_ABBR}
\title{IFC weekday abbreviations}
\format{
An object of class \code{character} of length 7.
}
\usage{
IFC_WDAY_ABBR
}
\description{
IFC weekday abbreviations
}
\keyword{datasets}
