% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\docType{data}
\name{IFC_WDAY_NAMES}
\alias{IFC_WDAY_NAMES}
\title{IFC weekday names (Sunday = 1, Saturday = 7)}
\format{
An object of class \code{character} of length 7.
}
\usage{
IFC_WDAY_NAMES
}
\description{
IFC weekday names (Sunday = 1, Saturday = 7)
}
\keyword{datasets}
