% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{ifc_year_day}
\alias{ifc_year_day}
\title{Create an IFC Year Day}
\usage{
ifc_year_day(year)
}
\arguments{
\item{year}{Integer. Calendar year.}
}
\value{
An \code{ifc_date} vector.
}
\description{
Year Day is the intercalary day at the end of each year. It is not part of
any month or week in the IFC system.
}
\examples{
ifc_year_day(2024)   # = Gregorian 2024-12-31
}
