% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiff_utils.R
\name{IFDtype}
\alias{IFDtype}
\title{IFD Type Detection}
\usage{
IFDtype(ifd)
}
\arguments{
\item{ifd}{an Image Field Directory as extracted by cpp_fastTAGS, cpp_getTAGS or subset of getIFD.}
}
\value{
an integer, with label attribute.
}
\description{
Detects IFD type
}
\keyword{internal}
