% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihs_transform.R
\name{ihs_transform}
\alias{ihs_transform}
\title{Inverse hyperbolic sine transformation of vector elements.}
\usage{
ihs_transform(x)
}
\arguments{
\item{x}{description a vector of numeric elements.}
}
\description{
\code{ihs_transform} returns the inverse hyperbolic sine transformation
}
\details{
This is a generic function that transforms a vector of elements using the inverse hyperbolic sine
transformation. This function will only take numeric atomic types such as doubles, integers, and floats.
}
\examples{
x <- c(1,2,3)
ihs_transform(x)

}
