% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_smoothLinLog.R
\name{inv_compLinLog}
\alias{inv_compLinLog}
\title{Inverse Compress LinLog Transformation}
\usage{
inv_compLinLog(x, hyper = 1000, base = 10, lin_size = 1)
}
\arguments{
\item{x}{A numeric vector.}

\item{hyper}{value where transition between Lin/Log is applied.}

\item{base}{base of Log scale.}

\item{lin_size}{size of the linear region equivalent as multiple of Log base unit. Default is 1.}
}
\value{
the inverse compLinLog transformation of the input.
}
\description{
Gets values back just to their original values before applying compLinLog.
}
