% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_utils.R
\name{inv_smoothAsinh}
\alias{inv_smoothAsinh}
\title{Inverse Smooth Asinh Transformation}
\usage{
inv_smoothAsinh(x, hyper = 1000)
}
\description{
Gets values back just to their original values before applying smoothAsinh
}
\keyword{internal}
