% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_utils.R
\name{keep_attributes}
\alias{keep_attributes}
\title{Attributes Keeper}
\usage{
keep_attributes(x, what)
}
\arguments{
\item{x}{list to pass to \code{'what'}.}

\item{what}{a build function.}
}
\value{
result of applying \code{'what'} on \code{'x'} with preserved attributes from \code{'x'}.
}
\description{
Helper to keep attributes while using IFC build functions.
}
\details{
all attributes of \code{'x'} will be used with the exception of "names", "dims", and "dimnames".
}
\keyword{internal}
