% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportToReport.R
\name{layoutReport}
\alias{layoutReport}
\title{Report Layout Extraction}
\usage{
layoutReport(graphs)
}
\arguments{
\item{graphs}{an `IFC_graphs` object extracted with features extracted.}
}
\value{
a list containing:\cr
-lay, a 3 columns (N, x, y) data.frame, where N is the graph index and x and y its coordinates on the layout,\cr
-mat, a matrix describing the layout.
}
\description{
Extracts report layout from `IFC_graphs` object.
}
\keyword{internal}
