% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log1_transform.R
\name{log1_transform}
\alias{log1_transform}
\title{Log(. + 1) transformation of vector elements.}
\usage{
log1_transform(x)
}
\arguments{
\item{x}{description a vector of numeric elements.}
}
\description{
\code{log1} returns the inverse hyperbolic sine transformation
}
\details{
This is a generic function that transforms a vector of elements using the log(. + 1)
transformation. This function will only take numeric atomic types such as doubles, integers, and floats.
}
