% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_table.R
\name{main_table}
\alias{main_table}
\title{Gives an output dataframe ready for kableExtra}
\usage{
main_table(..., last_panel)
}
\arguments{
\item{...}{fixest objects. This will be the regressions for however many fixest objects needed.}

\item{last_panel}{the final regression panel to put on. This will add on the fixed effects.}
}
\description{
\code{main_table} returns the main table broken down by subsets of fullsample/weekend/weekdays.
}
\details{
This function relies on dplyr, modelsummary, purrr, and tibble packages to work.
}
\examples{
main_table(full_sample, weekends, weekdays)

}
