% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_table_bootstrap.R
\name{main_table_bootstrap}
\alias{main_table_bootstrap}
\title{Gives an output dataframe ready for kableExtra when using fwildclusterboot standard errors}
\usage{
main_table_bootstrap(..., last_panel, display_fe = T)
}
\arguments{
\item{...}{fixest objects. This will be the regressions for however many fixest objects needed.}

\item{last_panel}{the final regression panel to put on. This will add on the fixed effects.}

\item{display_fe}{Whether or not to display the fixed effects chart in the table}
}
\description{
\code{main_table_bootstrap} returns the main table when using fwildcluster boot standard errors
}
\details{
This function relies on dplyr, modelsummary, purrr, and tibble packages to work.
}
\examples{
main_table_bootstrap(list(boostrap_1, boostrap_2), last_panel = list(boostrap_1, boostrap_2), display_fe = F)

}
