% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_table_semester.R
\name{main_table_semester}
\alias{main_table_semester}
\title{Finds the Average Length of IFC Moratoria.}
\usage{
main_table_semester(full_sample, weekends, weekdays)
}
\arguments{
\item{full_sample}{a list of fixest objects. This will be the regressions for the full sample.}

\item{weekends}{a list of fixest objects. This will be the regressions for the weekend sample.}

\item{weekdays}{a list of fixest objects. This will be the regressions for the weekdays.}
}
\description{
\code{main_table_semester} returns the main table broken down by subsets of fullsample/weekend/weekdays.
}
\details{
This function relies on dplyr, modelsummary, and tibble packages to work.
}
\examples{
main_table(full_sample, weekends, weekdays)

}
