% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_utils.R
\name{map_ansi}
\alias{map_ansi}
\title{ANSI Color Mapping}
\usage{
map_ansi(color, toANSI = TRUE)
}
\arguments{
\item{color}{a character vector Default is missing.}

\item{toANSI}{whether to convert color to ANSI or back. Default is \code{TRUE}.}
}
\value{
an integer vector when \code{'toANSI'} is \code{TRUE} or a character vector.
}
\description{
Converts R color to ANSI and inversely
}
\keyword{internal}
