% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_utils.R
\name{map_color}
\alias{map_color}
\title{Color Mapping}
\usage{
map_color(color, toR = TRUE)
}
\arguments{
\item{color}{a character vector Default is missing.}

\item{toR}{whether to convert \code{'color'} to R or back. Default is \code{TRUE}.}
}
\value{
a character vector.
}
\description{
Converts IDEAS/INSPIRE color to R and inversely
}
\keyword{internal}
