% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_utils.R
\name{map_set}
\alias{map_set}
\title{Set Mapping}
\usage{
map_set(x, to = TRUE, set1 = c(), set2 = c(), nomatch = FALSE)
}
\arguments{
\item{x}{a vector. Default is missing.}

\item{to}{whether to \code{'x'} from \code{'set1'} to \code{'set2'} or inversely. Default is \code{TRUE}.}

\item{set1, set2}{a vector. Default is \code{c()}.}

\item{nomatch}{whether no match should produce an error. Default is \code{FALSE}.}
}
\value{
a vector with values replacement.
}
\description{
Converts vector from one set of matching value to the other
}
\details{
only values found will be replaced; if not found, value is kept as is.
}
\keyword{internal}
