% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_utils.R
\name{map_style}
\alias{map_style}
\title{Style Mapping}
\usage{
map_style(style, toR = FALSE)
}
\arguments{
\item{style}{a pch (converted to integer) or a character vector. Default is missing.}

\item{toR}{whether to convert color toR or back. Default is FALSE.}
}
\value{
an integer vector when toR is TRUE or a character vector.
}
\description{
Converts IDEAS/INSPIRE style toR and inversely
}
\keyword{internal}
