% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{modifyGamma}
\alias{modifyGamma}
\title{Image Gamma Modification}
\usage{
modifyGamma(V, gamma = 1)
}
\arguments{
\item{V}{named NumericVector of channel display properties containing 'xmin', 'xmax', 'xmid' and 'ymid'.}

\item{gamma}{gamma}
}
\description{
Determines best xmid, ymid computes ymid for a given gamma
}
\keyword{internal}
