% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moratorium_lengths.R
\name{moratorium_lengths}
\alias{moratorium_lengths}
\title{Return a tibble of the moratorium lengths.}
\usage{
moratorium_lengths()
}
\description{
\code{moratorium_lengths} returns a tibble of moratorium lengths
}
\details{
This is a function that requires no input. Only returns the moratorium lengths and corresponding schools.
Must have the tidyverse package loaded in to do this.
}
\examples{
moratorium_lengths()

}
