% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moratorium_schools.R
\name{moratorium_schools}
\alias{moratorium_schools}
\title{Returns list of universities which underwent a moratorium from 2014-2019}
\usage{
moratorium_schools()
}
\description{
\code{moratorium_schools} returns a list of universities in which a moratorium has occurred due to death in addition to never-treated (but had death) schools.
}
\details{
This function simply returns a list of universities that went through moratorium. Does not rely on anything.
}
\examples{
moratorium_schools()

}
