% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{myScales}
\alias{myScales}
\title{LinLog Transformation for IFC Graphs Plotting Scales}
\usage{
myScales(x = list(), y = list())
}
\description{
Helper to rescale and label axes when linlog transformation is used
}
\keyword{internal}
