% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/never_treated_no_death.R
\name{never_treated_no_death}
\alias{never_treated_no_death}
\title{Returns list of universities which were never treated and never had a fraternity death from 2014-2019.}
\usage{
never_treated_no_death()
}
\description{
\code{never_treated_no_death} returns a list of universities in which are on the top 50 greek schools but did not have a death nor a moratorium.
}
\details{
This function simply returns a list of universities. Does not rely on anything.
}
\examples{
never_treated_no_death()

}
