% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_utils.R
\name{next_bool}
\alias{next_bool}
\title{Next Component Prediction}
\usage{
next_bool(x = "", count = 0L)
}
\arguments{
\item{x}{a string, current component.}

\item{count}{an integer, representing current number of opened/closed bracket.}
}
\value{
a vector of next allowed components.
}
\description{
Helper to define next allowed component in a boolean vector.
}
\keyword{internal}
