% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nschools.R
\name{nschools}
\alias{nschools}
\title{Counts Number of Unique Items in a Data Frame.}
\usage{
nschools(data, column_name)
}
\arguments{
\item{data}{a dataframe or tibble to pass in.}

\item{column_name}{the column you want to find the number of unique universities of.}
}
\description{
\code{nschools} returns the number of distinct values in a column. In this case, schools.
}
\details{
This is a function that takes the main data on schools, and counts how many schools there are by using
the distinct() function.
}
\examples{
nschools(dataframe, columnname)


}
