% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectColorize.R
\name{objectColorize}
\alias{objectColorize}
\title{Object Colorizer}
\usage{
objectColorize(mat, color)
}
\arguments{
\item{mat}{a [0,1] numeric matrix.}

\item{color}{a color}
}
\value{
a 3D array where 3rd dimension is rgb.
}
\description{
Colorizes a [0,1] grayscale image.
}
\keyword{internal}
