% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectResize.R
\name{objectResize}
\alias{objectResize}
\title{IFC_object Resizing}
\usage{
objectResize(
  mat,
  size = NULL,
  add_noise = TRUE,
  random_seed = 1,
  bg = 0,
  sd = 0
)
}
\arguments{
\item{mat}{a numeric matrix.}

\item{size}{a vector of final dimensions (Height, Width) of the desired returned mat. Default is NULL for no change.
if 'size' is of length one then only Width is changed.}

\item{add_noise}{if TRUE adds normal noise when size is larger than mat dimensions using rnorm(), from \pkg{Rcpp}. Default is TRUE.}

\item{random_seed}{a single value, interpreted as an integer when add_noise is set to TRUE. Default is 1.}

\item{bg}{mean value of the background added if add_noise is TRUE. Default is 0.}

\item{sd}{standard deviation of the background added if add_noise is TRUE. Default is 0.}
}
\value{
a resized matrix with padding background if desired size is larger than original mat dimensions.
}
\description{
Resizes mat to new dimensions.
}
