% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectWrite.R
\name{objectWrite}
\alias{objectWrite}
\title{Object File Export}
\usage{
objectWrite(x, type, ...)
}
\arguments{
\item{x}{a numeric matrix.}

\item{type}{image type. Supported values are: "bmp", "jpeg", "png", and "tiff".}

\item{...}{other arguments to be passed.}
}
\description{
Exports images to various types.
}
\keyword{internal}
