% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletteIFC.R
\name{paletteIFC}
\alias{paletteIFC}
\title{R/IDEAS Color Palette Mapping}
\usage{
paletteIFC(
  x = c("", "palette", "palette_R", "to_light", "to_dark")[1],
  col = "White"
)
}
\arguments{
\item{x}{either "", "palette","palette_R", to_light, to_dark. Default is "".}

\item{col}{a compatible color to transform to color or lightModeColor. Default is "White".\cr
if 'x' == to_light, function will convert 'col' to lightModeColor.\cr
if 'x' == to_dark, function will convert 'col' to color.\cr
if 'col' is not found or 'x' is anything else then a data.frame of compatible colors is returned.}
}
\value{
IFC palette of available colors.
}
\description{
Maps colors between IDEAS and R.
}
