% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{pan_hist}
\alias{pan_hist}
\title{Lattice Histogram Panel Contructor}
\source{
derived from \pkg{lattice} from Deepayan Sarkar
}
\usage{
pan_hist(
  x,
  type,
  br,
  normalize,
  fill,
  lwd,
  lty,
  col,
  alpha,
  ylim,
  bin,
  border,
  include.lowest = TRUE,
  right = TRUE
)
}
\description{
Helper to create histogram panel
}
\keyword{internal}
