% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{pan_key}
\alias{pan_key}
\title{Lattice Key Panel Contructor}
\source{
derived from \pkg{lattice} from Deepayan Sarkar
}
\usage{
pan_key(key, corner = c(0, 0.98), x = corner[1], y = corner[2])
}
\description{
Helper to add key to panel
}
\keyword{internal}
