% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{pan_smooth}
\alias{pan_smooth}
\title{Histogram Smooth Constructor}
\source{
derived from \pkg{lattice} from Deepayan Sarkar
}
\usage{
pan_smooth(
  x,
  type,
  br,
  normalize,
  fill,
  lwd,
  lty,
  col,
  alpha,
  ylim,
  bin,
  border,
  include.lowest = TRUE,
  right = TRUE,
  factor = 0
)
}
\description{
Helper to smooth histogram
}
\keyword{internal}
