% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_utils.R
\name{parseFCSname}
\alias{parseFCSname}
\title{FCS Name Parser}
\usage{
parseFCSname(x)
}
\arguments{
\item{x}{a character vector of FCS features names.}
}
\value{
a 2 columns data.frame of names (PnN) and alt-names (PnS).
}
\description{
Separates names and alt-names from FCS features names.
}
\details{
when created FCS features names are formatted as 'PnN < PnS >'.
The current function allows the separation between PnN and PnS.
}
\keyword{internal}
