% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{plot_backend}
\alias{plot_backend}
\title{`IFC_plot` Plotting}
\usage{
plot_backend(obj, backend = "lattice")
}
\arguments{
\item{obj}{an object of class `IFC_plot` as created by \code{\link{plotGraph}}.}

\item{backend}{the backend used for drawing `IFC_plot` see \code{\link{plotGraph}}.}
}
\value{
nothing.
}
\description{
Helper to plot `IFC_plot`
}
\keyword{internal}
