% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{plot_base}
\alias{plot_base}
\title{`IFC_plot` Conversion to 'base' Plot}
\usage{
plot_base(obj)
}
\arguments{
\item{obj}{an object of class `IFC_plot` as created by \code{\link{plotGraph}}.}
}
\description{
Helper to convert `IFC_plot` to 'base' plot.
}
\keyword{internal}
