% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{plot_base_error}
\alias{plot_base_error}
\title{Erroneous `IFC_plot` Conversion to 'base' Plot}
\usage{
plot_base_error(obj)
}
\arguments{
\item{obj}{an object of class `IFC_plot` as created by \code{\link{plotGraph}}.}
}
\value{
nothing, plot `IFC_plot` with 'base' backend.
}
\description{
Helper to produce 'base' plot with error message from `IFC_plot`.
}
\keyword{internal}
