% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{plot_default_args}
\alias{plot_default_args}
\title{`IFC_plot` Default Arguments Computation}
\usage{
plot_default_args(obj = list())
}
\arguments{
\item{obj}{an object of class `IFC_plot` as created by \code{\link{plotGraph}}.}
}
\value{
a list of default arguments for plotting.
}
\description{
Helper to extract default plot arguments from `IFC_plot`
}
\keyword{internal}
