% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{plot_lattice}
\alias{plot_lattice}
\title{`IFC_plot` Conversion to 'lattice' Plot}
\usage{
plot_lattice(obj)
}
\arguments{
\item{obj}{an object of class `IFC_plot` as created by \code{\link{plotGraph}}.}
}
\value{
a lattice trellis from `IFC_plot`.
}
\description{
Helper to convert `IFC_plot` to 'lattice' plot.
}
\keyword{internal}
