% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{plot_lattice_empty}
\alias{plot_lattice_empty}
\title{`IFC_plot` Lattice Empty Plot Creation}
\usage{
plot_lattice_empty(obj)
}
\arguments{
\item{obj}{an object of class `IFC_plot` as created by \code{\link{plotGraph}}.}
}
\value{
an empty lattice trellis from `IFC_plot`.
}
\description{
Helper to create an empty `IFC_plot`
}
\keyword{internal}
