% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{plot_lattice_error}
\alias{plot_lattice_error}
\title{Erroneous `IFC_plot` Conversion to 'lattice' Plot}
\usage{
plot_lattice_error(obj)
}
\arguments{
\item{obj}{an object of class `IFC_plot` as created by \code{\link{plotGraph}}.}
}
\value{
a lattice trellis from `IFC_plot`.
}
\description{
Helper to produce 'lattice' plot with error message from `IFC_plot`.
}
\keyword{internal}
