% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{plot_raster}
\alias{plot_raster}
\title{`IFC_plot` Conversion to 'raster' Plot}
\usage{
plot_raster(obj, pntsonedge = FALSE)
}
\arguments{
\item{obj}{an object of class `IFC_plot` as created by \code{\link{plotGraph}}.}

\item{pntsonedge}{whether points outside of plotting region should be bounded on the edge. Default is FALSE to clip points.
NA can be used to produce hybrid display, with plot being drawn with `pntsonedge` = FALSE on top of plot with `pntsonedge` = TRUE.}
}
\value{
plot `IFC_plot` with 'raster' backend and invisibly returns the computed raster.
}
\description{
Helper to convert `IFC_plot` to 'raster' plot.
}
\keyword{internal}
