% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{plot_raster_empty}
\alias{plot_raster_empty}
\title{`IFC_plot` Raster Empty Plot Creation}
\usage{
plot_raster_empty(obj)
}
\arguments{
\item{obj}{an object of class `IFC_plot` as created by \code{\link{plotGraph}}.}
}
\value{
nothing, plot `IFC_plot` with 'raster' backend.
}
\description{
Helper to create an empty `IFC_plot`
}
\keyword{internal}
