% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{plot_stats}
\alias{plot_stats}
\title{`IFC_plot` Statistics Extraction}
\usage{
plot_stats(obj)
}
\arguments{
\item{obj}{an object of class `IFC_plot` as created by \code{\link{plotGraph}}.}
}
\value{
a table of statistics.
}
\description{
Helper to extract `IFC_plot` statistics.
}
\keyword{internal}
