% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pnt_in_ell}
\alias{pnt_in_ell}
\title{Point in Ellipse}
\arguments{
\item{pnt}{NumericVector, x and y coordinates of the point.}

\item{ell}{NumericVector, coordinates of the ellipse.}
}
\description{
Checks if 2D point lies within a defined ellipse.
}
\keyword{internal}
