% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsGetAffiliation.R
\name{popsGetAffiliation}
\alias{popsGetAffiliation}
\title{IFC_pops Affiliation Finder}
\usage{
popsGetAffiliation(pops, operators = c("And", "Or", "Not", "(", ")"))
}
\arguments{
\item{pops}{list of populations}
}
\description{
Helper that extracts populations dependencies/affiliations.
}
\keyword{internal}
