% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsGetSiblings.R
\name{popsGetSiblings}
\alias{popsGetSiblings}
\title{IFC_pops Sibling Population Identification}
\usage{
popsGetSiblings(obj, pops)
}
\arguments{
\item{obj}{an `IFC_data` object extracted with features extracted.}

\item{pops}{graphical populations names to get siblings of.}
}
\value{
names of population siblings.
}
\description{
Gives names of graphical pops's siblings in a `IFC_data` object.
}
\keyword{internal}
