% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsGetSiblings.R
\name{popsGetSiblings1}
\alias{popsGetSiblings1}
\title{IFC_pops Sibling Population from Same Base Identification}
\usage{
popsGetSiblings1(obj, pops)
}
\arguments{
\item{obj}{an `IFC_data` object extracted with features extracted.}

\item{pops}{graphical populations names to get siblings of.}
}
\value{
names of population siblings.
}
\description{
Gives names of graphical pops's siblings in a `IFC_data` object.
Siblings are built from a same population but different regions drawn on axes
of same feature(s) with same transformation(s) applied if any.
}
\keyword{internal}
