% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsGetSiblings.R
\name{popsGetSiblings2}
\alias{popsGetSiblings2}
\title{IFC_pops Sibling Population from Same Region Identification}
\usage{
popsGetSiblings2(obj, pops)
}
\arguments{
\item{obj}{an `IFC_data` object extracted with features extracted.}

\item{pops}{graphical populations names to get siblings of.}
}
\value{
names of population siblings.
}
\description{
Gives names of graphical pops's siblings in a `IFC_data` object.
Siblings are built from different populations but a same region drawn on axes
of same feature(s) with same transformation(s) applied if any.
}
\keyword{internal}
