% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsOrderNodes.R
\name{popsOrderNodes}
\alias{popsOrderNodes}
\title{IFC_pops Population Level Ordering}
\usage{
popsOrderNodes(pops)
}
\arguments{
\item{pops}{list of populations.}
}
\description{
Helper that sort populations so that populations that depend from other are placed after.
}
\keyword{internal}
