% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsStyle.R
\name{popsStyle}
\alias{popsStyle}
\title{IFC_pops Style}
\usage{
popsStyle(obj)
}
\arguments{
\item{obj}{an `IFC_data` object extracted by ExtractFromDAF(extract_features = TRUE) or ExtractFromXIF(extract_features = TRUE).}
}
\value{
a named list of UTF-8 characters.
}
\description{
Helper to extract population style.
}
\keyword{internal}
