% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsTree.R
\name{popsTree}
\alias{popsTree}
\title{IFC_pops Tree}
\usage{
popsTree(pops)
}
\arguments{
\item{pops}{list of tree characters.}
}
\description{
Helper to extract hierarchy tree from pops.
}
\keyword{internal}
