% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsWithin.R
\name{popsWithin}
\alias{popsWithin}
\title{IFC_pops Checker for Object Membership to Populations}
\source{
For pnt_in_poly_algorithm, Trigonometry, is an adaptation of Jeremy VanDerWal's code \url{https://github.com/jjvanderwal/SDMTools}
}
\usage{
popsWithin(
  pops,
  regions,
  features,
  pnt_in_poly_algorithm = 1,
  pnt_in_poly_epsilon = 1e-12,
  display_progress = TRUE,
  title_progress = "",
  ...
)
}
\arguments{
\item{pops}{list of populations.}

\item{regions}{list of regions.}

\item{features}{dataframe of features.}

\item{pnt_in_poly_algorithm}{algorithm used to determine if object belongs to a polygon region or not. Default is 1.\cr
Note that for the moment only 1(Trigonometry) is available.}

\item{pnt_in_poly_epsilon}{epsilon to determine if object belongs to a polygon region or not. It only applies when algorithm is 1. Default is 1e-12.}

\item{display_progress}{whether to display a progress bar. Default is TRUE.}

\item{title_progress}{character string, giving the title of the progress bar. Default is "".}

\item{...}{other arguments to be passed.}
}
\description{
Helper that will be used by popsCompute to determine which objects are within populations or not.
}
\keyword{internal}
