% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsNetwork.R
\name{print.IFC_network}
\alias{print.IFC_network}
\title{Print Network}
\usage{
\method{print}{IFC_network}(x, ...)
}
\arguments{
\item{x}{object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print `IFC_network`
}
\keyword{internal}
