% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_utils.R
\name{protectn}
\alias{protectn}
\title{Name Protection}
\usage{
protectn(name)
}
\arguments{
\item{name}{population/region names}
}
\description{
Helper to protect population/region name.
}
\keyword{internal}
