% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_utils.R
\name{quad_coords}
\alias{quad_coords}
\title{Quad Coordinates}
\usage{
quad_coords(coords, part, xlim, ylim)
}
\arguments{
\item{coords}{a list containing x and y numeric vectors of at least 2 elements each, if not 1st element will be repeated twice.}

\item{part}{a string referencing sync part, as returned by \code{sync_part()}.}

\item{xlim}{x-limits. Default is missing to use \code{sync_lims()}, otherwise it will be \code{sync_lims(xlim)}.}

\item{ylim}{y-limits. Default is missing to use \code{sync_lims()}, otherwise it will be \code{sync_lims(ylim)}.}
}
\value{
coordinates for the corresponding \code{part}
}
\description{
Helper to construct coordinates for "quad" sync.
}
\details{
1st x and 1st y should be the x/y coordinates where parts 1-4 are separated.
Then, min/max xlim min/max ylim will be distributed to each part:\cr
-1: x = x[1], min(xlim); y = y[1], max(ylim)\cr
-2: x = x[1], max(xlim); y = y[1], max(ylim)\cr
-3: x = x[1], max(xlim); y = y[1], min(ylim)\cr
-4: x = x[1], min(xlim); y = y[1], min(ylim)\cr
otherwise, x = x[1], x[2]; y = y[1], y[2].\cr
In addition, \code{-Inf} x values will be replaced by \code{min(xlim)} and \code{+Inf} x values by \code{max(xlim)}, same for y values and \code{'ylim'}.
}
\keyword{internal}
