% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs.R
\name{readFCSdata}
\alias{readFCSdata}
\title{FCS Data Parser}
\usage{
readFCSdata(
  fileName,
  text,
  version,
  start = 0,
  end = 0,
  scale = TRUE,
  display_progress = TRUE,
  ...
)
}
\arguments{
\item{fileName}{path to file.}

\item{text}{a named list of keywords values.}

\item{version}{version of FCS specification.}

\item{start}{offset of text start. Default is 0.}

\item{end}{offset of text end. Default is 0.}

\item{scale}{whether to apply data scaling. It only applies when fcs file is stored as DATATYPE "I". Default is TRUE.\cr}

\item{display_progress}{whether to display a progress bar. Default is TRUE.}

\item{...}{other arguments to be passed.}
}
\description{
Helper to parse data segment from Flow Cytometry Standard (FCS) compliant files.
}
\keyword{internal}
