% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs.R
\name{readFCStext}
\alias{readFCStext}
\title{FCS Text Parser}
\usage{
readFCStext(
  fileName,
  delimiter,
  start = 0,
  end = 0,
  encoding = "UTF-8",
  empty = FALSE,
  trim = "none",
  ...
)
}
\arguments{
\item{fileName}{path to file.}

\item{delimiter}{delimiter used to separate keyword-value pairs. /!\ NOTE that files with 0x00 'delimiter' can _NOT_ be parsed.}

\item{start}{offset of text start. Default is 0.}

\item{end}{offset of text end. Default is 0.}

\item{encoding}{name of the encoding for raw to character conversion. Default is "UTF-8".}

\item{empty}{whether to allow empty values when parsing text segment. Default is FALSE.}

\item{trim}{remove whitespace in keywords names. Default is "none". Allowed are "both", "left", "right" and "none".}

\item{...}{other arguments to be passed.}
}
\description{
Helper to parse text segment from Flow Cytometry Standard (FCS) compliant files.
}
\keyword{internal}
