% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGatingStrategy.R
\name{readGatingStrategy}
\alias{readGatingStrategy}
\title{Gating Strategy File Reader}
\usage{
readGatingStrategy(fileName, ...)
}
\arguments{
\item{fileName}{path to file. It should be a .ast, .cif, .daf, .ist, .rif or .xml file.}

\item{...}{other arguments to be passed.}
}
\value{
A named list of class `IFC_gating`, whose members are:\cr
-description, a list of descriptive information,\cr
-pops, a list describing populations found,\cr
-regions, a list describing how regions are defined.
}
\description{
Extracts Gating Strategy from Files.
}
