% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redefine_utils.R
\name{redefine_features_def_feat}
\alias{redefine_features_def_feat}
\title{IFC_features_def Feature Redefinition}
\usage{
redefine_features_def_feat(
  features_def,
  to_match_feat = NULL,
  to_replace_feat = NULL,
  force_default = FALSE,
  ...
)
}
\arguments{
\item{features_def}{an `IFC_features_def` object or a list containing features definition. Default is missing.}

\item{to_match_feat}{a string with a features_def name to use for matching 'features_def' names. Default is NULL}

\item{to_replace_feat}{a string of features_def name to use for replacing 'features_def' names. Default is NULL}

\item{force_default}{whether to force default names for features (except the one defined by 'to_replace_feat').
This removes custom names and replaces them with default values. Default is FALSE.}

\item{...}{Other arguments to be passed.}
}
\value{
an `IFC_features_def` object, or a list containing features definition
}
\description{
Helper to rename a feature within IFC_features_def.
}
\keyword{internal}
