% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redefine_utils.R
\name{redefine_masks}
\alias{redefine_masks}
\title{IFC_masks Redefinition}
\usage{
redefine_masks(
  masks,
  images,
  new_images_names = images$name,
  to_match_mask = NULL,
  to_replace_mask = NULL,
  ...
)
}
\arguments{
\item{masks}{an `IFC_masks` object or a data.frame containing masks definition and name. Default is missing.}

\item{images}{a data.frame containing images definition. Default is missing.}

\item{new_images_names}{a vector of image name to use for replacing 'images' names. Default is images$name}

\item{to_match_mask}{a string with a mask name to use for matching 'masks' names. Default is NULL}

\item{to_replace_mask}{a string of mask name to use for replacing 'masks' names. Default is NULL}

\item{...}{Other arguments to be passed.}
}
\value{
a list whose members are:
-masks, an `IFC_masks` object or data.frame containing masks definition and name.
-images, a data.frame containing images definition.
}
\description{
Helper to modify features_def according to masks or images redefinition
}
\keyword{internal}
