% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redefine_utils.R
\name{redefine_masks_image}
\alias{redefine_masks_image}
\title{IFC_masks Image Redefinition}
\usage{
redefine_masks_image(masks, images, new_images_names = images$name, ...)
}
\arguments{
\item{masks}{an `IFC_masks` object or a data.frame containing masks definition and name. Default is missing.}

\item{images}{a data.frame containing images definition. Default is missing.}

\item{new_images_names}{a vector of image name to use for replacing 'images' names. Default is images$name}

\item{...}{Other arguments to be passed.}
}
\value{
a list whose members are:
-masks, an `IFC_masks` object or data.frame containing masks definition and name.
-images, a data.frame containing images definition.
}
\description{
Helper to rename images within masks definition.
}
\keyword{internal}
