% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redefine_utils.R
\name{redefine_obj}
\alias{redefine_obj}
\title{IFC_data Redefinition}
\usage{
redefine_obj(obj, new_feat_def)
}
\arguments{
\item{obj}{an `IFC_data` object. Default is missing.}

\item{new_feat_def}{a list with new definitions from modify_features().}

\item{...}{Other arguments to be passed.}
}
\value{
an `IFC_data` object.
}
\description{
Helper to modify images, masks, features in a IFC data object when features_def is changed
}
\keyword{internal}
