% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reghdfe_pois.R
\name{reghdfe_pois}
\alias{reghdfe_pois}
\title{High dimension fixed effects regression (Poisson) with fixest package.}
\usage{
reghdfe_pois(data, outcome, explanatory_vars, fixed_effects, cluster)
}
\arguments{
\item{data}{tibble or dataframe.}

\item{outcome}{outcome of interest. Must be a string.}

\item{explanatory_vars}{vector of explanatory variables. Must be a string.}

\item{fixed_effects}{a vector of fixed effects. Must be a string.}

\item{cluster}{column to cluster by. Must be a string.}
}
\description{
\code{reghdfe_pois} returns the fixest object consisting of the regression equation.
}
\details{
This function calls the fixest::fepois() function, but simiplifies the typing.
}
\examples{
reghdfe(mtcars, "mpg", "hp", "am", "am")

}
