% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_utils.R
\name{remove_ext}
\alias{remove_ext}
\title{File Extenstion Removal}
\usage{
remove_ext(x)
}
\arguments{
\item{x}{a file path}
}
\description{
Remove file extension from file path
}
\details{
file extension will be removed
}
\keyword{internal}
