% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{scale_trans}
\alias{scale_trans}
\title{LinLog Transformation of Ticks and Labels}
\usage{
scale_trans(hyper = 1000, base = 10, lin_comp = log(base))
}
\description{
Helper to rescale and label axes when linlog transformation is used
}
\keyword{internal}
