% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_utils.R
\name{setloc}
\alias{setloc}
\title{Set Locale}
\usage{
setloc(locale = getloc())
}
\arguments{
\item{locale}{a named vector with locale values to modify. Default is getloc().}
}
\value{
a named vector of the locale values used before the application of the function.
}
\description{
Sets aspects of locale
}
\keyword{internal}
