% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_utils.R
\name{setPB}
\alias{setPB}
\title{Progress Bar Updater}
\usage{
setPB(pb, value = NULL, title = NULL, label = NULL)
}
\arguments{
\item{pb}{an object of class `IFC_progress` containing a progress bar of class `txtProgressBar`, `winProgressBar` or `Progress`.}

\item{value}{new value for the progress bar.}

\item{title, label}{character strings, giving the 'title'(='message' for shiny progress bar) and the 'label'(='detail' for shiny progress bar).}
}
\description{
Updates a progress bar.
}
\keyword{internal}
