% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writetiff.R
\name{setupimage}
\alias{setupimage}
\title{Image Setup for TIFF Export}
\usage{
setupimage(image, what, compression, endianness = .Platform$endian)
}
\arguments{
\item{image}{a matrix or array or representing the image:\cr
-2D matrix [h,w] will be written as single grayscale [h,w] IFD (Image Field Directory),\cr
-3D array [h,w,c] will be written as single IFD of multichannel [h,w,c],\cr
-4D array [h,w,c,f] will be written as multiple (multiframe) IFDs of multichannel [h,w,c],\cr
with h=height, w=width, c=channel, f=frame.\cr
Eventually, a raw vector with `dims`, `what` and `comp` attributes can be directly passed asis (allowing the use \code{compression} other than deflate).}

\item{what}{bits mode used to store image. Allowed are \code{"uint8"}, \code{"int8"}, \code{"uint16"}, \code{"int16"}, \code{"uint32"}, \code{"int32"}, \code{"float"} and \code{"double"}.}

\item{compression}{whether image should be lossless compressed with deflate algorithm. Default is FALSE.}

\item{endianness}{The endian-ness ("big" or "little") of the return object. Default is .Platform$endian.\cr}
}
\value{
It returns a raw vector with `dims`, `what` and `comp` attributes.
}
\description{
Perpares image for TIFF export.
}
\keyword{internal}
