% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{/.ifc_interval}
\alias{/.ifc_interval}
\title{Divide an IFC interval by a time unit}
\usage{
\method{/}{ifc_interval}(e1, e2)
}
\arguments{
\item{e1}{An \code{ifc_interval}.}

\item{e2}{A string: \code{"day"}, \code{"week"}, or \code{"month"}.}
}
\value{
A numeric vector of the span expressed in the given unit.
}
\description{
Divide an IFC interval by a time unit
}
