% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_utils.R
\name{specialr}
\alias{specialr}
\title{Special Character Replacement}
\usage{
specialr(
  string = "",
  replacement = "_",
  specials = "\\\\\\\\|\\\\/|\\\\:|\\\\*|\\\\?|\\\\\\"|\\\\<|\\\\>|\\\\|"
)
}
\arguments{
\item{string}{string where specials will be replaced if found.}

\item{replacement}{string replacement. Default is "_".}

\item{specials}{Default is '\\\\|\\/|\\:|\\*|\\?|\\"|\\<|\\>|\\|'.}
}
\description{
Helper to replace special character.
}
\keyword{internal}
