% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_utils.R
\name{splitf}
\alias{splitf}
\title{File Path Decomposition}
\usage{
splitf(file = NULL)
}
\arguments{
\item{file}{path to file}
}
\value{
a named vector with chunks of 'file'\cr
dir: full path directory of 'file'\cr
parent: first parent directory of 'file'\cr
ext: 'file' extension without leading dot\cr
short: 'file' with no extension nor dir\cr
input: 'file' path as it was provided.
}
\description{
Helper that will split file name into chunks
}
\keyword{internal}
