% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_utils.R
\name{splitn}
\alias{splitn}
\title{String Decomposition with Operators}
\usage{
splitn(
  definition,
  all_names,
  operators = c("And", "Or", "Not", "(", ")"),
  splitter = "[`~splitter~`]"
)
}
\arguments{
\item{definition}{population definition to be splitted}

\item{all_names}{the names of all allowed populations}

\item{operators}{operators used. Default is c("And", "Or", "Not", "(", ")").}

\item{splitter}{the splitter that will be used to help splitting}
}
\description{
Helper that will split population definition into chunks of names and operators.
}
\keyword{internal}
