% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_utils.R
\name{splitp}
\alias{splitp}
\title{String Decomposition with Placeholders}
\usage{
splitp(write_to = "\%d/\%s_fromR.\%e")
}
\arguments{
\item{write_to}{string. Default is "\%d/\%s_fromR.\%e"}
}
\description{
Helper aiming to detect placeholder pattern
}
\details{
-\%s: shortname (i.e. basename without extension)\cr
-\%p: first parent directory\cr
-\%d: full path directory\cr
-\%e: file extension\cr
-\%o: object id\cr
-\%c: channel
}
\keyword{internal}
