% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_utils.R
\name{splits}
\alias{splits}
\title{Sync Name Decomposition}
\usage{
splits(x, name = TRUE, split = "|")
}
\arguments{
\item{x}{string}

\item{name}{whether to return name only. Default is \code{TRUE}.}

\item{split}{string used for splitting. Default is \code{"|"}.}
}
\value{
when \code{'x'} is of length 0, \code{""} is returned. Otherwise, the \code{'name'} when \code{'name'} is \code{TRUE} or the \code{'split'} non-containing remaining string when \code{FALSE}.
}
\description{
Helper that will split text into chunks.
}
\details{
\code{'x'} is supposed to be a string composed of a \code{'name'} followed by \code{'split'} and ending with a string that does not contain \code{'split'}.
}
\keyword{internal}
