% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{statsCompute}
\alias{statsCompute}
\title{IFC_stats Computation}
\usage{
statsCompute(
  obj,
  stats,
  width = 80 * length(stats),
  height = 240,
  xlocation = 0,
  ylocation = 0,
  ...
)
}
\arguments{
\item{obj}{an `IFC_data` object}

\item{stats}{list of statistics. It will be coerced by buildStats()}

\item{width}{desired width. Default is 80 * (1 + length(stats)).}

\item{height}{desired height Default is 400.}

\item{xlocation}{x location in analysis workspace. Default is 0.}

\item{ylocation}{y location in analysis workspace. Default is 0.}

\item{...}{Other arguments to be passed.}
}
\value{
an object of class `IFC_pops`.
}
\description{
Function used to compute `IFC_stats` object\cr
}
\keyword{internal}
