% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{StatsReport}
\alias{StatsReport}
\title{Statistical Report Generation}
\usage{
StatsReport(obj, stats)
}
\arguments{
\item{obj}{an `IFC_data` object.}

\item{stats}{template defining stats to compute as extracted by getSTATSREPORT().}
}
\value{
a named vector with extracted statistics.
}
\description{
Generates stats report from `IFC_data` object.
}
\keyword{internal}
