% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redefine_utils.R
\name{swap_channel}
\alias{swap_channel}
\title{Channel Swap}
\usage{
swap_channel(obj, chan1, chan2, BF = TRUE, MODE = 1)
}
\arguments{
\item{obj}{an `IFC_data` object extracted with features extracted.}

\item{chan1, chan2}{an integer index of channel. 'chan1' and 'chan2' should be different.}

\item{BF}{should 'from' channel be considered as brightfield. Default is TRUE.}

\item{MODE}{collection mode (as retrieved by getInfo) determining the range. Default is 1.}
}
\value{
a list, intended to be passed to 'new_feat_def' argument of getFromNamespace("redefine_obj", "IFC") whose members are:\cr
-features_def, an 'IFC_features_def' object, or a list containing features definition\cr
-masks, an 'IFC_masks' object or a data.frame containing masks definition and name.\cr
-images, a data.frame containing images definition.
}
\description{
Swaps Channels within `IFC_data` object
}
\details{
'BF' and 'MODE' will be used only if 'chan1' or 'chan2' is not found in 'obj'.\cr
/!\ NOTE: In case of conflict between resulting names, 'chan2' will be preferred, maximizing number of features with 'chan2'.
}
\keyword{internal}
