% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_utils.R
\name{sync_check}
\alias{sync_check}
\title{Sync Check}
\usage{
sync_check(regions)
}
\arguments{
\item{regions}{an `IFC_regions` object.}
}
\value{
data.frame with \code{'name'} of invalid sync attribute and \code{'error'} description
}
\description{
Helper to check sync attribute.
}
\keyword{internal}
