% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_utils.R
\name{sync_clip}
\alias{sync_clip}
\title{Sync Clip}
\usage{
sync_clip(x, lims)
}
\arguments{
\item{x}{a numeric vector}

\item{lims}{limits. Default is missing to use \code{sync_lims()}, otherwise it will be \code{sync_lims(lims)}.}
}
\description{
Helper to clip vector to sync limits.
}
\details{
\code{'x'} will be clip to \code{'lims'} excluding \code{'lims'}
}
\keyword{internal}
