% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_utils.R
\name{sync_coords}
\alias{sync_coords}
\title{Sync Coordinates}
\usage{
sync_coords(type, coords, part, xlim, ylim)
}
\arguments{
\item{type}{type of sync. Expecting \code{"dual"} or \code{"quad"}.}

\item{coords}{a list containing x and y numeric vectors of at least 2 elements each, if not 1st element will be repeated twice.}

\item{part}{a string referencing sync part, as returned by \code{sync_part()}.}

\item{xlim}{x-limits. Default is missing to use \code{sync_lims()}, otherwise it will be \code{sync_lims(xlim)}.}

\item{ylim}{y-limits. Default is missing to use \code{sync_lims()}, otherwise it will be \code{sync_lims(ylim)}.}
}
\value{
coordinates for the corresponding \code{part}
}
\description{
Helper to construct coordinates for sync.
}
\details{
see \code{dual_coords()} or \code{quad_coords()}.
}
\keyword{internal}
