% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_utils.R
\name{sync_create}
\alias{sync_create}
\title{Create Sync}
\usage{
sync_create(reg, xlim, ylim, forbidden = character())
}
\arguments{
\item{reg}{a pre built region see \code{\link{buildRegion}} except that \code{reg$type} should be \code{"dual"} or \code{"quad"} instead of "line" or "rect" respectively.}

\item{xlim}{x-limits. Default is missing to use \code{sync_lims()}, otherwise it will be \code{sync_lims(xlim)}.}

\item{ylim}{y-limits. Default is missing to use \code{sync_lims()}, otherwise it will be \code{sync_lims(ylim)}.}

\item{forbidden}{a character vector of already given sync names, see \code{sync_name()}. Default is \code{character()}.}
}
\value{
a list of region(s)
}
\description{
Helper to construct sync regions.
}
\details{
if \code{reg$type} is not \code{"dual"} or \code{"quad"}, it is returned as list(reg).
/!\ Note that when \code{'xlim'} is missing or is \code{c(-Inf,Inf)} label position \code{'reg$cx'} is hard to define so it falls back to \code{'reg$x'}. Same for \code{'ylim'} and \code{'reg$cy'} for \code{"quad"}.
}
\keyword{internal}
