% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_utils.R
\name{sync_lims}
\alias{sync_lims}
\title{Sync Limits}
\usage{
sync_lims(lims = c(-Inf, Inf))
}
\arguments{
\item{lims}{limits. Default is \code{c(-Inf,Inf)}.}
}
\description{
Sync Limits.
}
\keyword{internal}
