% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_utils.R
\name{sync_name}
\alias{sync_name}
\title{Sync Name}
\usage{
sync_name(reg, what)
}
\arguments{
\item{reg}{a region, though it can be any R object}

\item{what}{a string. If provided \code{sync_type(reg)} should be identical to \code{'what'}, otherwise \code{""} is returned.}
}
\value{
a string containing sync name
}
\description{
Helper to retrieve sync name.
}
\details{
sync name is the result of splitting \code{attr(reg, "sync")} with \code{"|"} and keeping all that is before the last \code{"|"} encountered.\cr
Note that \code{NA_character_} will be returned in case of error during splitting i.e. no \code{"|"} found or \code{length(attr(reg, "sync")) != 1}.
}
\keyword{internal}
