% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_utils.R
\name{sync_parse}
\alias{sync_parse}
\title{Sync Parse}
\usage{
sync_parse(regions, what)
}
\arguments{
\item{regions}{an `IFC_regions` object, though it can be any R object}

\item{what}{a string}
}
\value{
a character vector of sync name.
}
\description{
Helper to serialize sync name.
}
\keyword{internal}
