% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_utils.R
\name{sync_part}
\alias{sync_part}
\title{Sync Part}
\usage{
sync_part(reg, what)
}
\arguments{
\item{reg}{a region, though it can be any R object}

\item{what}{a string. If provided \code{sync_type(reg)} should be identical to \code{'what'}, otherwise \code{""} is returned.}
}
\value{
a string containing sync part.
}
\description{
Helper to retrieve sync part.
}
\details{
sync part is the result of retrieving number after the last \code{"|"} encountered in \code{attr(reg, "sync")}.
}
\keyword{internal}
