% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_utils.R
\name{sync_type}
\alias{sync_type}
\title{Sync Type}
\usage{
sync_type(reg, what)
}
\arguments{
\item{reg}{a region, though it can be any R object}

\item{what}{a string. If provided returned sync type should be identical to \code{'what'}, otherwise \code{""} is returned.}
}
\value{
a string containing sync type.
}
\description{
Helper to retrieve sync type.
}
\details{
sync type is the result of \code{substr(attr(reg, "sync"), 1, 4)}.
In addition, when this result is \code{"dual"} or \code{"quad"} and \code{'what'} is provided, \code{attr(reg, "sync")}:\cr
-should end with \code{"|1"} or \code{"|2"} for \code{"dual"}, otherwise \code{""} is returned.\cr
-should end with \code{"|1"}, \code{"|2"}, \code{"|3"} or \code{"|4"} for \code{"quad"} otherwise \code{""} is returned.
}
\keyword{internal}
