% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_utils.R
\name{sync_validate}
\alias{sync_validate}
\title{Sync Validate}
\usage{
sync_validate(regions, partial = FALSE, quietly = TRUE, caller = character(0))
}
\arguments{
\item{regions}{an `IFC_regions` object.}

\item{partial}{a bool whether number of regions and their coordinates match sync attribute. Default is \code{FALSE} to only validate type.}

\item{quietly}{a bool whether to silent warning when a sync attribute is not valid. Default is \code{TRUE}.}

\item{caller}{a string, name of the calling function. Default is \code{character(0)}.}
}
\value{
an `IFC_regions` with sync validated attributes
}
\description{
Helper to validate sync regions.
}
\keyword{internal}
