% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_panels.R
\name{table_panels}
\alias{table_panels}
\title{Gets a dataframe of modelsummary tables pasted.}
\usage{
table_panels(...)
}
\arguments{
\item{...}{as many models as you like!}
}
\description{
\code{table_panels} returns a table of pasted together data frames.
}
\details{
This function relies on dplyr, modelsummary, and tibble packages to work.
}
\examples{
table_panels(full_sample, weekends, weekdays)

}
