% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiff_utils.R
\name{testXIF}
\alias{testXIF}
\title{RIF/CIF Image Order Test}
\usage{
testXIF(fileName)
}
\arguments{
\item{fileName}{path of file.}
}
\value{
an integer\cr
-1: not a XIF file\cr
0: non regular XIF file, i.e. no mask found after 1st Image itself after 1st IFD\cr
+1: regular XIF file, i.e. a mask is found after 1st Image itself after 1st IFD.
}
\description{
Tests order of IFD within RIF and XIF file
}
\keyword{internal}
