% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texttomatrix.R
\name{texttomatrix}
\alias{texttomatrix}
\title{Text to Image Conversion}
\usage{
texttomatrix(text)
}
\arguments{
\item{text}{a character string.}
}
\description{
Converts string to matrix of 0 and 1.
}
\details{
at symbol @ is a special character used for scale
}
\keyword{internal}
