% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{to_list_node}
\alias{to_list_node}
\title{XML Node to List Conversion}
\usage{
to_list_node(x, max = 5)
}
\arguments{
\item{x}{A document, node, or node set.}

\item{max}{maximum number of recurrence into subnodes. Default is 5.}
}
\description{
Helper to convert xml node to R list.
}
\details{
it acts as_list but value returned is different, with attributes
expanded to sublists rather than recovered as attributes
}
\keyword{internal}
