% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{to_xml_list}
\alias{to_xml_list}
\title{List to XML Node Conversion}
\usage{
to_xml_list(x, name, kids, indent = "  ", escape = "\\n")
}
\arguments{
\item{x}{a list to convert}

\item{name}{name of the node to create}

\item{kids}{a list containing children xml nodes elements (each elements should come from to_xml_list)}

\item{indent}{indent used for kids when provided. Default is "  ".}

\item{escape}{escape used for kids when provided. Default is "\\n".}
}
\description{
Helper to convert R list to xml node (character representation).
}
\details{
it acts as_list but value returned is different, with attributes
expanded to sublists rather than recovered as attributes
}
\keyword{internal}
