% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_utils.R
\name{toCapFirstOnly}
\alias{toCapFirstOnly}
\title{First Letter Only Capitalization}
\usage{
toCapFirstOnly(text)
}
\arguments{
\item{text}{a string}
}
\description{
Helper to capitalize the first letter of strings and leave the rest to lower case
}
\keyword{internal}
