% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{toEllipse}
\alias{toEllipse}
\title{Ellipsoid Polygon Constructor}
\usage{
toEllipse(gate, theta = 2 * pi, npoints = 100)
}
\arguments{
\item{gate}{list containing x and y ellipse boundaries coordinates.}

\item{theta}{rotation angle of the ellipse. Default is 2*pi. It should not be modified since ellipse gate are axis-aligned.}

\item{npoints}{number of polygon vertices desired.}
}
\description{
Helper to transform gate boundaries to ellipsoid polygon.
}
\keyword{internal}
