% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingML.R
\name{toXML2_boolpop_gs}
\alias{toXML2_boolpop_gs}
\title{Boolean Population GatingML Conversion to XML2}
\usage{
toXML2_boolpop_gs(obj, pop)
}
\arguments{
\item{obj}{an `IFC_data` object.}

\item{pop}{a member of `IFC_pops` object.}
}
\value{
a list of xml_node.
}
\description{
Helper to convert boolean population to XML nodes in GatingML files.
}
\keyword{internal}
