% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toXML2_features.R
\name{toXML2_features}
\alias{toXML2_features}
\title{IFC_features XML Conversion}
\usage{
toXML2_features(
  features,
  verbose = FALSE,
  display_progress = TRUE,
  title_progress = ""
)
}
\arguments{
\item{features}{an `IFC_features` object.}

\item{verbose}{whether to display message about current action. Default is \code{FALSE}.}

\item{display_progress}{whether to display a progress bar. Default is \code{TRUE}.}

\item{title_progress}{character string, giving the title of the progress bar. Default is \code{""}.}
}
\value{
a xml_node.
}
\description{
Helper to convert features (`IFC_features` object) to XML nodes.
}
\keyword{internal}
