% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toXML2_features_def.R
\name{toXML2_features_def}
\alias{toXML2_features_def}
\title{IFC_features_def XML Conversion}
\usage{
toXML2_features_def(features_def, verbose = verbose)
}
\arguments{
\item{verbose}{whether to display message about current action. Default is FALSE.}

\item{features}{an `IFC_features_def` object.}
}
\value{
a XML::xmlNode.
}
\description{
Helper to convert features definition (`IFC_features_def` object) to XML nodes.
}
\keyword{internal}
