% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingML.R
\name{toXML2_gating}
\alias{toXML2_gating}
\title{GatingML Conversion}
\usage{
toXML2_gating(pop, reg, verbose = FALSE)
}
\arguments{
\item{pop}{a member of `IFC_pops` object.}

\item{reg}{a member of `IFC_regions` object.}

\item{verbose}{whether to display message about current action. Default is FALSE.}
}
\value{
a xml_node.
}
\description{
Helper to convert pops and regions to XML nodes in GatingML files.
}
\keyword{internal}
