% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toXML2_graphs.R
\name{toXML2_graphs}
\alias{toXML2_graphs}
\title{IFC_graphs XML Conversion}
\usage{
toXML2_graphs(graphs, verbose = FALSE)
}
\arguments{
\item{graphs}{an `IFC_graphs` object.}

\item{verbose}{whether to display message about current action. Default is FALSE.}
}
\value{
a xml_node.
}
\description{
Helper to convert graphs (`IFC_graphs` object) to XML nodes.
}
\keyword{internal}
