% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingML.R
\name{toXML2_graphs_gs}
\alias{toXML2_graphs_gs}
\title{IFC_graphs GatingML Conversion}
\usage{
toXML2_graphs_gs(graphs)
}
\arguments{
\item{graphs}{an `IFC_graphs` object.}
}
\value{
a xml_node.
}
\description{
Helper to convert graphs (`IFC_graphs` object) to XML nodes in GatingML files.
}
\keyword{internal}
