% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toXML2_images.R
\name{toXML2_images}
\alias{toXML2_images}
\title{IFC_images XML Conversion}
\usage{
toXML2_images(
  images,
  verbose = FALSE,
  display_progress = TRUE,
  title_progress = ""
)
}
\arguments{
\item{images}{an `IFC_images` object.}

\item{verbose}{whether to display message about current action. Default is FALSE.}

\item{title_progress}{character string, giving the title of the progress bar. Default is "".}
}
\value{
a xml_node.
}
\description{
Helper to convert images (`IFC_images` object) to XML nodes.
}
\keyword{internal}
