% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toXML2_masks.R
\name{toXML2_masks}
\alias{toXML2_masks}
\title{IFC_masks XML Conversion}
\usage{
toXML2_masks(masks, verbose = FALSE)
}
\arguments{
\item{masks}{an `IFC_masks` object.}

\item{verbose}{whether to display message about current action. Default is FALSE.}
}
\value{
a xml_node.
}
\description{
Helper to convert masks (`IFC_masks` object) to XML nodes.
}
\keyword{internal}
