% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toXML2_pops.R
\name{toXML2_pops}
\alias{toXML2_pops}
\title{IFC_pops XML Conversion}
\usage{
toXML2_pops(
  pops,
  verbose = FALSE,
  display_progress = TRUE,
  title_progress = "",
  ...
)
}
\arguments{
\item{pops}{an `IFC_pops` object.}

\item{verbose}{whether to display message about current action. Default is FALSE.}

\item{display_progress}{whether to display a progress bar. Default is TRUE.}

\item{title_progress}{character string, giving the title of the progress bar. Default is "".}

\item{...}{other arguments to be passed.}
}
\value{
a xml_node.
}
\description{
Helper to convert populations (`IFC_pops` object) to XML nodes.
}
\keyword{internal}
