% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toXML2_regions.R
\name{toXML2_regions}
\alias{toXML2_regions}
\title{IFC_regions XML Conversion}
\usage{
toXML2_regions(regions, verbose = FALSE)
}
\arguments{
\item{regions}{an `IFC_regions` object.}

\item{verbose}{whether to display message about current action. Default is FALSE.}
}
\value{
a xml_node.
}
\description{
Helper to convert regions (`IFC_regions` object) to XML nodes.
}
\keyword{internal}
