% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingML.R
\name{toXML2_spillover_gs}
\alias{toXML2_spillover_gs}
\title{Spillover GatingML Conversion}
\usage{
toXML2_spillover_gs(spillover, name)
}
\arguments{
\item{spillover}{a spillover matrix. It has to have colnames and rownames.}

\item{name}{the name to use to identify the node.}
}
\value{
a xml_node.
}
\description{
Helper to convert spillover to XML nodes in GatingML files.
}
\keyword{internal}
