% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{toXML2_stats}
\alias{toXML2_stats}
\title{IFC_stats XML Conversion}
\usage{
toXML2_stats(stats, verbose = FALSE)
}
\arguments{
\item{stats}{an `IFC_stats` object.}

\item{verbose}{whether to display message about current action. Default is FALSE.}
}
\value{
a xml_node.
}
\description{
Helper to convert stats (`IFC_stats` object) to XML nodes.
}
\keyword{internal}
