% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends_test.R
\name{trends_test}
\alias{trends_test}
\title{Creates the data needed to perform the random trends test.}
\usage{
trends_test(data, window, cluster_blocks)
}
\arguments{
\item{data}{an object/dataframe/tibble.}

\item{window}{how big the event window is (in weeks)}

\item{cluster_blocks}{a number for how many days you want clustered together. I usually do 7 for week.}
}
\description{
\code{trends_test} returns a tibble that is ready for the random trends test.
}
\details{
This function creates the event study tibble needed to do analysis in event study framework.
}
\examples{
trends_test(daily_crime,  window = 8, cluster_blocks = 7)

}
