% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_utils.R
\name{trunc_string}
\alias{trunc_string}
\title{String Truncation}
\usage{
trunc_string(x, n = 22)
}
\arguments{
\item{x}{a string}

\item{n}{desired length}
}
\description{
Truncs character strings
}
\details{
x will be truncated according to 'n' parameter. If x is longer than n '...' are appended.
}
\keyword{internal}
