% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{type_constr}
\alias{type_constr}
\title{Histogram Type Constructor}
\source{
derived from \pkg{lattice} from Deepayan Sarkar
}
\usage{
type_constr(x, br, type, include.lowest = TRUE, right = TRUE, plot = FALSE)
}
\description{
Helper to construct histogram
}
\keyword{internal}
