% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redefine_utils.R
\name{usedefault_obj}
\alias{usedefault_obj}
\title{IFC_data Default Naming}
\usage{
usedefault_obj(obj, ...)
}
\arguments{
\item{obj}{an `IFC_data` object. Default is missing.}

\item{...}{Other arguments to be passed.}
}
\value{
an `IFC_data` object.
}
\description{
Helper to reset masks and features names to their default values in a `IFC_data` object.
}
\keyword{internal}
