% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{val_constr}
\alias{val_constr}
\title{Histogram Val Constructor}
\source{
derived from \pkg{lattice} from Deepayan Sarkar
}
\usage{
val_constr(x, h, type)
}
\description{
Helper to construct histogram
}
\keyword{internal}
