% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{wrap_label}
\alias{wrap_label}
\title{Graphical Text Wrap}
\usage{
wrap_label(label, split = "\\\\s", max_width = NULL, ...)
}
\arguments{
\item{label}{a character string.}

\item{split}{see \code{split} \code{\link[base]{strsplit}} called with fixed = FALSE.}

\item{max_width}{maximum width in npc to occupy, leave \code{NULL} to auto-compute.}

\item{...}{other arguments to be passed to \code{\link[grid]{textGrob}}.}
}
\value{
a line wrapped label
}
\description{
Wraps text grob label into multiple lines
}
\keyword{internal}
