% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs.R
\name{writeFCS}
\alias{writeFCS}
\title{FCS File Writer}
\usage{
writeFCS(
  data,
  write_to,
  overwrite = FALSE,
  delimiter = "/",
  cytometer = "Image Stream",
  keywords = list(NULL)
)
}
\arguments{
\item{data}{a data.frame}

\item{write_to}{path of exported file. Note that file extension has to be a .fcs.}

\item{overwrite}{whether to overwrite file or not. Default is \code{FALSE}.
Note that if \code{TRUE}, it will overwrite exported file, unless it is an original source file (i.e. not written by \pkg{IFC}).\cr}

\item{delimiter}{an ASCII character to separate the FCS keyword-value pairs. Default is \code{"/"}.}

\item{cytometer}{string, if provided it will be used to fill '$CYT' keyword.Default is \code{"Image Stream"}.}

\item{keywords}{keyword-value pairs. Default is \code{list(NULL)}.}
}
\value{
invisibly returns full path of exported file.
}
\description{
Writes data to a Flow Cytometry Standard (FCS) file.
}
\keyword{internal}
