% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiff_utils.R
\name{writeIFD}
\alias{writeIFD}
\title{Image Field Directory Writer}
\usage{
writeIFD(
  ifd,
  r_con,
  w_con,
  pos = 0,
  extra = NULL,
  endianness = .Platform$endian,
  ...
)
}
\arguments{
\item{ifd}{an ifd extracted by cpp_fastTAGS}

\item{r_con}{a connection opened for reading}

\item{w_con}{a connection opened for writing}

\item{pos}{current position within 'w_con'. Default is 0.}

\item{extra}{extra entries to add to 'ifd'. Default is NULL}

\item{endianness}{the desired endian-ness ("big" or "little"). Default is .Platform$endian.\cr
Endianness describes the bytes order of data stored within the files. This parameter may not be modified.}
}
\value{
the position within 'w_con' after 'IFD' and 'extra' content have been written\cr
}
\description{
Writes Image Field Directory (IFD)
}
\keyword{internal}
