% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XIFtoTIFF.R
\name{XIFtoTIFF}
\alias{XIFtoTIFF}
\title{RIF/CIF File Conversion to TIFF}
\usage{
XIFtoTIFF(
  fileName,
  write_to,
  objects,
  offsets,
  fast = TRUE,
  bits = 8,
  endianness = .Platform$endian,
  verbose = FALSE,
  verbosity = 1,
  overwrite = FALSE,
  display_progress = TRUE,
  ...
)
}
\arguments{
\item{fileName}{path of file to subset.
It has to be a '.rif' or '.cif' file.}

\item{write_to}{pattern used to export file.
Placeholders, like "\%d/\%s_fromR.\%e", will be substituted:\cr
-\%d: with full path directory of 'fileName'\cr
-\%p: with first parent directory of 'fileName'\cr
-\%e: with extension of 'fileName' (without leading .)\cr
-\%s: with shortname from 'fileName' (i.e. basename without extension).\cr
Exported file extension has to be .tif or .tiff.}

\item{objects}{integer vector, IDEAS objects ids numbers to use. If missing, the default, all objects will be used.}

\item{offsets}{object of class `IFC_offset`. If missing, the default, offsets will be extracted from 'fileName'.\cr
This param is not mandatory but it may allow to save time for repeated XIF export on same file.}

\item{fast}{whether to fast extract 'objects' or not. Default is TRUE.
Meaning that 'objects' will be extracting expecting that 'objects' are stored in ascending order.\cr
Note that a warning will be sent if an 'object' is found at an unexpected order.
In such a case you may need to rerun function with 'fast' = FALSE.
If set to FALSE, all raw object_ids will be scanned from 'fileName' to ensure extraction of desired 'objects'.\cr
IMPORTANT: whatever this argument is, features are extracted assuming an ascending order of storage in file.}

\item{bits}{gray levels depth. Default is 8 (i.e. 2^8 = [0,255] gray levels). Allowed are 8 and 16.}

\item{endianness}{The endian-ness ("big" or "little") of the target system for the file. Default is .Platform$endian.\cr
Endianness describes the bytes order of data stored within the files. This parameter may not be modified.}

\item{verbose}{whether to display information (use for debugging purpose). Default is FALSE.}

\item{verbosity}{quantity of information displayed when verbose is TRUE; 1: normal, 2: rich. Default is 1.}

\item{overwrite}{whether to overwrite file or not. Default is FALSE.\cr
Note that if TRUE, it will overwrite exported file if path of 'fileName' and deduced from 'write_to' arguments are different.
Otherwise, you will get an error saying that overwriting source file is not allowed.\cr
Note also that an original file, i.e. generated by IDEAS(R) or INSPIRE(R), will never be overwritten.
Otherwise, you will get an error saying that overwriting original file is not allowed.\cr}

\item{display_progress}{whether to display a progress bar. Default is TRUE.}

\item{...}{other arguments to be passed.}
}
\value{
It invisibly returns full path of exported file.
}
\description{
Converts and subsets RIF or CIF files to TIFF.
This function is experimental.
}
\details{
This function is experiment and under development inputs and outputs may change in the future
}
\keyword{internal}
