% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{xml_new_node}
\alias{xml_new_node}
\title{List to XML Conversion}
\usage{
xml_new_node(name, attrs, .children, text, ...)
}
\arguments{
\item{name}{name of the node to create.}

\item{attrs}{a named list of name-value pairs to be used as attributes for the XML node.}

\item{.children}{a list containing XML node elements or content.}

\item{text}{the text content for the new XML node.}
}
\value{
an R object that points to the C-level structure instance.
}
\description{
Helper to convert R list to xml node (character representation).
}
\keyword{internal}
