% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addScaleBar.R
\name{addScaleBar}
\alias{addScaleBar}
\title{Image Scale Bar Incrustation}
\usage{
addScaleBar(
  image,
  size = 7,
  style = c("dash", "line")[1],
  color = "white",
  pix = c(0.3, 0.5, 1)[1],
  xoff = 0,
  yoff = 0
)
}
\arguments{
\item{image}{a [0,1] image.}

\item{size}{positive integer. Scale's bar size in micro-meter. Default is \code{7}.}

\item{style}{a character string. Scale's bar style, either \code{"dash"} or \code{"line"}. Default is \code{"dash"}.}

\item{color}{a character string. Color of the scale. Default is \code{"white"} .}

\item{pix}{positive numeric. Size of one pixel in micro-meter. Default is \code{0.3} for 60x magnification, use \code{0.5} for 40x and \code{1} for 20x.}

\item{xoff}{x offset in image to draw scale, starting from bottom left corner. Default is \code{0}.}

\item{yoff}{y offset in image to draw scale, starting from bottom left corner. Default is \code{0}.}
}
\value{
an image with scale added to the bottom left corner.
}
\description{
Adds scale bar to image
}
\keyword{internal}
