% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{adjustGraph}
\alias{adjustGraph}
\title{IFC Graph Adjustment}
\usage{
adjustGraph(obj, graph, adjust_graph = TRUE, ...)
}
\arguments{
\item{obj}{an object of class `IFC_data` extracted by ExtractFromDAF(extract_features = TRUE) or ExtractFromXIF(extract_features = TRUE).}

\item{graph}{a graph from 'obj' or a list that can be coerced by \code{\link{buildGraph}}.}

\item{adjust_graph}{whether to try to adjust graph(s) when possible. Default is TRUE.\cr
-TRUE, graph(s) will be kept if possible using only regions, pops it depends that can be found in 'obj',\cr
-FALSE, graph(s) will be kept only if all features, regions, pops it refers to are found in 'obj',\cr
-NA, graph(s) will be removed no matter if features, regions, pops it refers to are found in 'obj'.}

\item{...}{other arguments to be passed.}
}
\description{
Helper to readjust `IFC_data` graphs in case of missing feature, region, population.
}
\keyword{internal}
