% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{adjustGraph}
\alias{adjustGraph}
\title{IFC Graph Adjustment}
\usage{
adjustGraph(obj, selection, list.only = TRUE, ...)
}
\arguments{
\item{obj}{an object of class `IFC_data` extracted by ExtractFromDAF(extract_features = TRUE) or ExtractFromXIF(extract_features = TRUE).}

\item{selection}{when provided, indices of desired graphs.\cr
Note that indices are read from left to right, from top to bottom.}

\item{...}{other arguments to be passed.}

\item{list_only}{whether to return a vector of elements that will be impacted by the removal. Default is TRUE.
If FALSE then modified object will be returned}
}
\description{
Helper to readjust `IFC_data` graphs in case of missing feature, region, population.
}
\keyword{internal}
