% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_utils.R
\name{applyTrans}
\alias{applyTrans}
\title{Apply Transformation}
\usage{
applyTrans(x, trans, inverse = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{trans}{the object returned by parseTrans().}

\item{inverse}{whether or not to apply the inverse transformation. Default is FALSE.}
}
\value{
the transformation of the input
}
\description{
Helper apply transformation.
}
\details{
for the moment, in addition to no transformation, only smoothLinLog and asinh are supported.
}
\examples{
x <- 10^(1:10)
## parse the transformation instruction
trans <- parseTrans("smoothLinLog|100|4")
## transform x
xx <- applyTrans(x, trans)
## inverse transform the transformed values
x_back <- applyTrans(xx, trans, inverse = TRUE)
stopifnot(all.equal(x, x_back))
}
\keyword{internal}
