% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base64_encode.R
\name{base64_encode}
\alias{base64_encode}
\title{Raw Images to Base64 Encoding}
\source{
\url{https://en.wikibooks.org/wiki/Algorithm_Implementation/Miscellaneous/Base64}
}
\usage{
base64_encode(x)
}
\arguments{
\item{x}{a raw vector.}
}
\description{
Encodes raw image vector to base64 string
}
\keyword{internal}
