% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{buildStats}
\alias{buildStats}
\title{IFC Statistics Coercion}
\usage{
buildStats(type, title, def, ...)
}
\arguments{
\item{type}{stats's type. Allowed are "COUNT","PERCENT_GATED","CONCENTRATION","PERCENT","MEAN","MEDIAN","STDDEV","MAD","CV","MINIMUM","MAXIMUM","GEOMETRIC_MEAN","MODE","VARIANCE","NAN","MEAN_RD","MEDIAN_RD".}

\item{title}{stats's title. If missing, it will be determined thanks to def.}

\item{def}{definition of the stats.}

\item{...}{Other arguments to be passed.}
}
\value{
a list containing all statistics information.
}
\description{
Helper to build a list to allow statistics export.
}
\details{
when 'type' is:\cr
- "COUNT","PERCENT_GATED","CONCENTRATION", 'def' will be "",
- "PERCENT", 'def' has to be a population name,
- "MEAN_RD","MEDIAN_RD" 'def' has to be the concatenation of a feature name and a population name collapse with "|". e.g. "Area_MC|All",
- otherwise, 'def' has to be a feature name.
}
\keyword{internal}
