% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checksumXIF.R
\name{checksumXIF}
\alias{checksumXIF}
\title{RIF/CIF File Checksum}
\usage{
checksumXIF(fileName)
}
\arguments{
\item{fileName}{path to file.}
}
\description{
This function returns XIF checksum..
}
\keyword{internal}
