% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compensation.R
\name{CompensateFromRIF}
\alias{CompensateFromRIF}
\title{Shorcut for Batch RIF Images Spatial Correction and Compensation}
\usage{
CompensateFromRIF(..., objects, offsets, spillover, display_progress = TRUE)
}
\arguments{
\item{...}{other arguments to be passed.}

\item{objects}{integer vector, IDEAS objects ids numbers to use.
This argument is not mandatory, if missing, the default, all objects will be used.}

\item{offsets}{object of class `IFC_offset`. 
This argument is not mandatory but it may allow to save time for repeated image export on same file.}

\item{spillover}{spillover matrix to apply. Default is missing to apply no compensation.
In such case, you will get images with spatial correction only. However, when set to NULL images 
will be compensated using spillover matrix set during acquisition. If not NULL the provided spillover 
matrix will be used to compensate images.}

\item{display_progress}{whether to display a progress bar. Default is TRUE.}
}
\value{
A list of matrices/arrays of images corresponding to objects extracted.
}
\description{
Function to shortcut extraction and RIF file spatial correction and compensation
}
\details{
this function is experimental inputs and outputs may change in the future.
}
\keyword{internal}
