% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_align}
\alias{cpp_align}
\title{Spatial Offsets Image Correction}
\arguments{
\item{mat, }{a NumericMatrix.}

\item{dx, }{a double x spatial offset. It has to be within ]-1,+1[.}

\item{dy, }{a double y spatial offset. It has to be within ]-1,+1[.}
}
\value{
a NumericMatrix.
}
\description{
This function uses bilinear interpolation to apply spatial offset correction on image
}
\details{
It is intended to be applied on raw images matrices from .rif files so has to generate spatial offset corrected image matrices.\cr
See William E. Ortyn et al. Sensitivity Measurement and Compensation in Spectral Imaging. Cytometry A 69 852-862 (2006).
\url{https://onlinelibrary.wiley.com/doi/full/10.1002/cyto.a.20306}
}
\keyword{internal}
