% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_cleanse}
\alias{cpp_cleanse}
\title{Matrix Cleanser}
\arguments{
\item{mat}{a NumericMatrix.}

\item{msk}{a IntegerMatrix.}

\item{add_noise}{logical, if true adds normal noise.
Rcpp::Rf_rnorm(bg, sd) function is used. Default is true.}

\item{bg}{double, mean value of the background added if add_noise is true. Default is 0.0}

\item{sd}{double, standard deviation of the background added if add_noise is true. Default is 0.0}
}
\value{
a NumericMatrix with replaced according to msk.
}
\description{
Replaces values in matrix mat according to mask msk.
Depending of add_noise parameter values will be replaced with noise or not.
}
\keyword{internal}
