% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_crop}
\alias{cpp_crop}
\title{Matrix Cropping}
\arguments{
\item{mat}{a numeric matrix.}

\item{new_height}{an unsigned integer, giving the new height of returned mat. Default is 0 for no change.}

\item{new_width}{an unsigned integer, giving the new width of returned mat. Default is 0 for no change.}
}
\value{
a cropped matrix.
}
\description{
Crops mat according to new_height and new_width parameters.
}
\keyword{internal}
