% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_extract}
\alias{cpp_extract}
\title{IFC_object Extraction}
\arguments{
\item{fname}{string, path to file}

\item{ifd}{List, ifd information of class IFC_ifd}

\item{colors}{List of colors to use.}

\item{channels}{DataFrame, channels information}

\item{chan_to_extract}{IntegerVector, channels to extract}

\item{extract_msk}{uint8_t, type of masked to extract.\cr
- 0: no mask\cr
- 1: at least one clipped\cr
- 2: at least one masked\cr
- 3: at least one MC}

\item{mode}{string, color mode export. Either "rgb", "gray" or "raw". Default is "raw".}

\item{size}{a length 2 IntegerVector of final dimensions (height,width) of the image. Default is 0,0 for no change.\cr}

\item{verbose}{bool, whether to display information (use for debugging purpose). Default is false.}
}
\description{
Extracts object from ifd
}
\keyword{internal}
