% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_fastTAGS}
\alias{cpp_fastTAGS}
\title{IFD Fast Tags Extraction}
\source{
TIFF 6.0 specifications archived from web \url{https://web.archive.org/web/20211209104854/https://www.adobe.io/open/standards/TIFF.html}
}
\arguments{
\item{fname}{string, path to file.}

\item{offset}{uint32_t, position of the IFD beginning.}

\item{verbose}{bool, whether to display information (use for debugging purpose). Default is 'false'.}
}
\description{
Returns TAGS contained within an IFD (Image Field Directory) entry.
}
\keyword{internal}
