% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_getTAGS}
\alias{cpp_getTAGS}
\title{IFD Tags Extraction}
\source{
TIFF 6.0 specifications available at \url{https://www.adobe.io/open/standards/TIFF.html}
}
\arguments{
\item{fname}{string, path to file.}

\item{offset}{uint32_t, position of the IFD beginning.}

\item{verbose}{bool, whether to display information (use for debugging purpose). Default is 'false'.}

\item{trunc_bytes}{uint8_t, number of bytes to extract for bytes/strings TAGS (1, 2, 6 or 7). Default is 22.}

\item{force_trunc}{bool, whether to force truncation for all TAGS types. Default is 'false'.}
}
\description{
Returns TAGS contained within an IFD (Image Field Directory) entry.
}
\keyword{internal}
