% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_mark}
\alias{cpp_mark}
\title{Matrix to Matrix Writer According to Mask with Offsets}
\arguments{
\item{A}{a NumericMatrix.}

\item{B}{a NumericMatrix.}

\item{mask}{a NumericMatrix.}

\item{xoff}{x offset in A to start writing B.}

\item{yoff}{x offset in A to start writing B.}

\item{invert}{a logical. Default is false.
When false, the default, values of B are written into A when mask is not 0.
When true, values of 1-B are written into A when mask is not 0.}
}
\description{
Writes matrix B in matrix A according to mask.
}
\keyword{internal}
