% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_mask}
\alias{cpp_mask}
\title{Equal Sized Matrix to Matrix Writer According to Mask}
\arguments{
\item{A}{a NumericMatrix.}

\item{B}{a NumericMatrix.}

\item{mask}{a NumericMatrix.}
}
\description{
Writes matrix B in matrix A according to mask.
If mask is not 0 B is written, A otherwise.
}
\keyword{internal}
