% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_num_to_string}
\alias{cpp_num_to_string}
\title{Numeric to String Formatting}
\arguments{
\item{x}{NumericVector.}

\item{precision, }{number of significant decimal digits to keep when x < 1. Default is 16.}
}
\value{
a string vector.
}
\description{
Formats numeric to string used for features, images, ... values conversion when exporting to xml.
}
\keyword{internal}
