% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_smoothLinLog}
\alias{cpp_smoothLinLog}
\title{Smooth LinLog Transformation with Rcpp}
\arguments{
\item{x}{NumericVector.}

\item{hyper}{double, value where transition between Lin/Log is applied.}

\item{base}{double, base of Log scale.}

\item{lin_comp}{double, value that is used to smooth transition between Lin/Log.}
}
\description{
Takes a numeric vector and return its transformation:
- to linear, if abs(x) < hyper.
- to log, if abs(x) > hyper.
}
\keyword{internal}
