% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{densCols}
\alias{densCols}
\title{Colors for Smooth Density Plots}
\source{
derived from \pkg{grDevices} R Core Team, Florian Hahne at FHCRC, originally
}
\usage{
densCols(
  x,
  y = NULL,
  nbin = 128,
  bandwidth,
  colramp = colorRampPalette(c("blue", "green", "red")),
  transformation = function(x) (x^0.125)
)
}
\description{
Helper to map density to colors
}
\keyword{internal}
