% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportToReport.R
\name{DisplayReport}
\alias{DisplayReport}
\title{Graphical and Statistic Report Display}
\usage{
DisplayReport(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{ExportToReport}} where write_to is inoperative.}
}
\value{
It invisibly returns NULL
}
\description{
Displays report from `IFC_data` object.
}
\keyword{internal}
