% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_study_day.R
\name{event_study_day}
\alias{event_study_day}
\title{Creates the Event Study tibble for analysis at Daily Level Data.}
\usage{
event_study_day(data, window, cluster_blocks, reference_days)
}
\arguments{
\item{data}{an object/dataframe/tibble.}

\item{window}{how big the event window is (in weeks)}

\item{cluster_blocks}{a number for how many days you want clustered together. I usually do 7 for week.}

\item{reference_days}{a number for which year you want to be the reference week It will be paste0("treatment_minus_", reference_year)}
}
\description{
\code{event_study_day} returns a tibble ready for event study analysis in units of days
}
\details{
This function creates the event study tibble needed to do analysis in event study framework.
}
\examples{
event_study_day(daily_crime,  window = 8, cluster_blocks = 7, reference_days =1 )

}
