% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractMasks_toMatrix.R
\name{ExtractMasks_toMatrix}
\alias{ExtractMasks_toMatrix}
\title{Shorcut for Batch Masks Extraction to Matrices/Arrays}
\usage{
ExtractMasks_toMatrix(
  fileName,
  display,
  offsets,
  objects,
  display_progress = TRUE,
  ...
)
}
\arguments{
\item{fileName}{path of file to read data from. This parameter can't be missing.}

\item{display}{object of class IFC_display, rich information extracted by \code{\link{getDisplayInfo}}. If missing, the default, 'display' will be extracted from fileName.\cr
This param is not mandatory but it may allow to pass custom display info by manually modifying display$Images for example.}

\item{offsets}{object of class IFC_offsets. If missing, the default, 'offsets' will be extracted from fileName.\cr
This param is not mandatory but it may allow to save time for repeated image export on same file.}

\item{objects}{integers, indices of objects to use.
This param is not mandatory, if missing, the default, all objects will be used.}

\item{display_progress}{whether to display a progress bar. Default is TRUE.}

\item{...}{other arguments to be passed to \code{\link{objectExtract}} with the exception of 'ifd' and 'export'(="matrix").\cr
If 'display' and/or 'offsets' are not provided arguments can also be passed to \code{\link{getDisplayInfo}} and/or \code{\link{getOffsets}}.}
}
\value{
A list of matrices/arrays of masks corresponding to objects extracted.
}
\description{
Function to shortcut extraction, normalization and eventually colorization of masks to matrix ! excludes image.
}
\details{
arguments of \code{\link{objectExtract}} will be deduced from \code{\link{ExtractMasks_toMatrix}} input arguments.
}
