% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs.R
\name{FCS_check_keywords}
\alias{FCS_check_keywords}
\title{FCS Keyword Checker}
\usage{
FCS_check_keywords(
  text,
  delimiter,
  version = 3,
  encoding = "UTF-8",
  fun = "warning",
  ...
)
}
\arguments{
\item{text}{a named list of keywords values.}

\item{delimiter}{delimiter used to separate keyword-value pairs. /!\ NOTE that files with 0x00 'delimiter' can _NOT_ be parsed.}

\item{version}{version to check keywords compliance against. Default is 3.0.}

\item{encoding}{name of the encoding for raw to character conversion. Default is "UTF-8".}

\item{fun}{function to execute when mandatory parameters are not met. Default is "warning". Allowed are "stop","warning","message","return".}

\item{...}{other arguments to be passed.}
}
\description{
Helper to check that FCS keyword-value pairs are compliant with specifications
}
\keyword{internal}
