% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_utils.R
\name{gen_altnames}
\alias{gen_altnames}
\title{Alternative Names Generator}
\usage{
gen_altnames(x, n = 10, forbidden = character())
}
\arguments{
\item{x}{a character vector.}

\item{n}{number of characters of the desired returned name. Default is 10.}

\item{forbidden}{forbidden character vector. Default is character().}
}
\value{
a character vector.
}
\description{
Generates unique non matching alternative names
}
\details{
'forbidden' should not encompass all possible returned value otherwise the function will never end.
}
\keyword{internal.}
