% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFullTag.R
\name{getFullTag}
\alias{getFullTag}
\title{Image Field Directory Full Tag Getter}
\source{
TIFF 6.0 specifications available at \url{https://www.adobe.io/open/standards/TIFF.html}
}
\usage{
getFullTag(fileName, ifd, tag)
}
\arguments{
\item{fileName}{path to file..}

\item{ifd}{only one sub element of IFD data information extracted by \code{\link{getIFD}}.}

\item{tag}{numeric value or character value of the desired tag present ifd.}
}
\value{
If tag is of type 2, it returns a character string.\cr
If tag is of type 1, 6 or 7 it returns a raw vector.\cr
For other types, it will return mapped value corresponding to TIFF 6.0 specifications.
}
\description{
Retrieves full tag value from IFDs (Image Field Directory) extracted by \code{\link{getIFD}}.
}
\details{
It is usefull to extract all information contained in a specific ifd tag since \code{\link{getIFD}} is designed to be run with argument trunc_bytes so as to only extract essential bytes to run faster and save memory.
Nonetheless, thanks to \code{\link{getFullTag}} users will still be able to get full extraction of specific tag.
}
