% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{getSTATSREPORT}
\alias{getSTATSREPORT}
\title{Statistics Report Template Extraction}
\usage{
getSTATSREPORT(fileName)
}
\arguments{
\item{fileName}{path to file..}

\item{...}{other arguments to be passed.}
}
\value{
a 6-columns character matrix describing report instructions:\cr
- 'name', for the desired name of exported 'statistics',\cr
- 'type', for the type of stats to return (either "value" or "ratio"),\cr
- 'population1', determines the population on which 'statistics' will be performed,\cr
- 'population2', determines the reference population (when 'type' is "ratio", see Details),\cr
- 'feature', determines the feature's name on which 'statistics' will be computed,\cr
- 'statistics', controls the mathematical function that will be applied (See Details).
}
\description{
Retrieves statistics report template from .ast / .daf files.
}
\details{
Allowed "statistics" names are: "Count","Mean","%Total","%Gated","%","Objects/mL","RD - Mean",
"Median","CV","stddev","NaN","MAD","min","RD - Median","Variance","max","geomean","Mode".\cr
For "%Total","%Gated","%","RD - Mean","RD - Median", "type" has to be ratio and both 'population1'
and 'population2' should be provided. Otherwise, 'type' is "value" and only 'population1' is mandatory.\cr
/!\ Note that "Mode" and "Objects/mL" can't be determined and will result in NA.
}
\keyword{internal}
