% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_utils.R
\name{hist_constr}
\alias{hist_constr}
\title{Histogram Constructor}
\usage{
hist_constr(x, breaks, include.lowest = TRUE, right = TRUE, ...)
}
\description{
Helper to construct histogram
}
\keyword{internal}
