% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_smoothLinLog.R
\name{inv_smoothLinLog}
\alias{inv_smoothLinLog}
\title{Inverse Smooth LinLog Transformation}
\usage{
inv_smoothLinLog(x, hyper = 1000, base = 10, lin_comp = log(base))
}
\arguments{
\item{x}{A numeric vector.}

\item{hyper}{value where transition between Lin/Log is applied.}

\item{base}{base of Log scale.}

\item{lin_comp}{value that is used to smooth transition between Lin/Log. Default is log(base).}
}
\value{
the inverse smoothLinLog transformation of the input.
}
\description{
Gets values back just to their original values before applying smoothLinLog.
}
