% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_utils.R
\name{map_color}
\alias{map_color}
\title{Color Mapping}
\usage{
map_color(color, toR = TRUE)
}
\arguments{
\item{color}{a string. Default is missing.}

\item{toR}{whether to convert color toR or back. Default is TRUE.}
}
\value{
a character string
}
\description{
Converts IDEAS/INSPIRE colors toR and inversely
}
\keyword{internal}
