% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_utils.R
\name{next_bool}
\alias{next_bool}
\title{Next Component Prediction}
\usage{
next_bool(x = character(0), count = 0L, obj_alias = character(0))
}
\arguments{
\item{x}{a string, current component. Default is character(0).}

\item{count}{an integer, representing current number of opened/closed bracket.}

\item{obj_alias}{a character vector, alias used for name(s). Default is character(0).}
}
\value{
a vector of next allowed components.
}
\description{
Helper to define next allowed component in a boolean vector.
}
\keyword{internal}
