% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectParam.R
\name{objectParam}
\alias{objectParam}
\title{IFC_object Object Extraction Parameters Definition}
\usage{
objectParam(
  display,
  composite = "",
  selection = "all",
  size = c(0, 0),
  force_width = TRUE,
  random_seed = NULL,
  removal = "none",
  add_noise = TRUE,
  full_range = FALSE,
  force_range = FALSE,
  bypass = FALSE,
  ...
)
}
\arguments{
\item{display}{object of class IFC_display, rich information extracted by \code{\link{getDisplayInfo}}. THis parameter can't be missing.}

\item{composite}{character vector of image composite. Default is "", for no image composite.\cr
Should be like "1.05/2.4/4.55" for a composition of 5 perc. of channel 1, 40 perc. of channel 2 and 50 perc. of channel 55.\cr
Note that channels should have been acquired and final image composition should be 100 perc., otherwise an error is thrown.\cr
Note that each composite will be appended after 'selection'.}

\item{selection}{physical channels to extract.\cr
Note that this parameter will be ordered.\cr
Default is "all" to extract all acquired channels.\cr
Use "none" to only extract composite.}

\item{size}{a length 2 integer vector of final dimensions of the image, height 1st and width 2nd. Default is c(0,0) for no change.}

\item{force_width}{whether to use information in display to fill size. Default is TRUE.
When set to TRUE, width of 'size' argument will be overwritten.}

\item{random_seed}{a single value, interpreted as an integer when 'add_noise' is set to TRUE. Default is NULL.}

\item{removal}{removal method: Either "none", "clipped", "masked", "MC".\cr
-"none", to keep image as is\cr
-"clipped", to remove clipped object from image.\cr
-"masked", to only keep masked object from image.\cr
-"MC", to only keep MC masked object from image.
This parameter will be repeated with rep_len() from \pkg{base} for every physical channel that needs to be extracted according to 'selection' and 'composite' parameters.}

\item{add_noise}{if TRUE adds normal noise to background using rnorm(), from \pkg{Rcpp}. Default is TRUE.\cr
Note that it is better to set it to FALSE when 'removal' is "masked" or "MC". Doing so will allow to place masked object in a zero filled background,
otherwise background will still be filled with noise.
This parameter will be repeated with rep_len() from \pkg{base} for every physical channel that needs to be extracted according to 'selection' and 'composite' parameters.}

\item{full_range}{only apply when mode is not "raw", if full_range is TRUE, then [0,4095] range will be kept. Default is FALSE.\cr
It is like "raw" mode but allowing normalization to [0,1].
This parameter will be repeated with rep_len() from \pkg{base} for every physical channel that needs to be extracted according to 'selection' and 'composite' parameters.\cr}

\item{force_range}{only apply when mode is not "raw", if force_range is TRUE, then image display range will be adjusted for each object resulting in normalization. Default is FALSE.\cr
Note that this parameter takes the precedence over 'full_range'.\cr
This parameter will be repeated with rep_len() from \pkg{base} for every physical channel that needs to be extracted according to 'selection' and 'composite' parameters.}

\item{bypass}{logical to avoid several checking. Default is FALSE.}

\item{...}{other arguments to be passed.}
}
\value{
A object of class IFC_param.
}
\description{
Defines IFC_object object extraction parameters.
}
