% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_utils.R
\name{parseTrans}
\alias{parseTrans}
\title{Parse Transformation Information}
\usage{
parseTrans(string)
}
\arguments{
\item{string}{character string describing transformation used and its hyperparameter.}
}
\value{
a list with 2 members what (=the transformation function) and args (=the parameters to pass to this function)
}
\description{
Helper to parse transformation parameter.
}
\details{
-If string is "P" no transformation will be applied.\cr
-If string is of length 1 and coercible to a numeric, smoothLinLog will be applied with hyper parameter set with as.numeric(string) value.\cr
-Otherwise, string will be split with "|", 1st element will be considered as function to call and all other elements will be passed after being coerced to numeric to this function in the order they are provided with the exception of 1st one; if coercion results in NA, the argument will be skipped.
}
\keyword{internal}
