% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractToFuns.R
\name{polyExtractTo}
\alias{polyExtractTo}
\title{Shortcut for Batch Images or Masks Extraction}
\usage{
polyExtractTo(
  ...,
  objects,
  offsets,
  display_progress = TRUE,
  image_type = "img",
  export = "matrix",
  mode = "raw"
)
}
\arguments{
\item{\dots}{arguments to be passed to \code{\link{objectExtract}} with the exception of \code{ifd} and \code{bypass}(=\strong{TRUE}).\cr
\strong{/!\\} If not any of '\code{fileName}', '\code{info}' and '\code{param}' can be found in \code{\dots} then \code{attr(offsets, "fileName_image")} will be used as '\code{fileName}' input parameter to pass to \code{\link{objectParam}}.}

\item{objects}{integer vector, IDEAS objects ids numbers to use.
This argument is not mandatory, if missing, the default, all objects will be used.}

\item{offsets}{object of class `IFC_offset`.
This argument is not mandatory but it may allow to save time for repeated image export on same file.\cr
If '\code{offsets}' are not provided extra arguments can also be passed with \code{\dots} to \code{\link{getOffsets}}.}

\item{display_progress}{whether to display a progress bar. Default is TRUE.}

\item{image_type}{(\code{\link{subsetOffsets}} argument) image_type of desired offsets. Default is "img". Allowed are "img" or "msk".}

\item{export}{(\code{\link{objectParam}} argument) format mode export. Either "file", "matrix", "base64". Default is "matrix".}

\item{mode}{(\code{\link{objectParam}} argument) color mode export. Either "raw", "rgb", "gray". Default is "raw".}
}
\value{
A list (for every extracted objects) of list (for every exported channels) depending on "export" parameter:\cr
-"matrix", a matrix when '\code{param$mode}' is set to "raw" or "gray" OR an array when '\code{param$mode}' == "rgb",\cr
-"base64", a data-uri string,\cr
-"file", an invisible file path corresponding to the location of exported file(s).
}
\description{
Function to shortcut extraction, normalization and eventually colorization of images or masks to various format.
}
\details{
arguments of \code{\link{objectExtract}} will be deduced from input arguments.\cr
If '\code{param}' is provided 'in \code{\dots}, \code{param$export}(=\strong{"base64"}) \strong{only} will be overwritten.
}
\keyword{internal}
