% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsGetLevels.R
\name{popsGetLevels}
\alias{popsGetLevels}
\title{IFC_pops Levels Dependency Determination}
\usage{
popsGetLevels(pops)
}
\arguments{
\item{pops}{list of populations.}
}
\description{
Helper that extract population levels.
}
\keyword{internal}
