% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsRename.R
\name{popsRename}
\alias{popsRename}
\title{Populations Renaming}
\usage{
popsRename(
  obj,
  old_names = character(),
  new_names = character(),
  loops = 10L,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{obj}{an `IFC_data`.}

\item{old_names}{character vector of name(s) of population(s) to rename inside 'obj'. Default is character().}

\item{new_names}{character vector of desired new population(s) name(s). Default is character().}

\item{loops}{a positive integer specifying the maximum number of recursive loops before raising an error. Default is 10L.}

\item{verbose}{whether to show a final message about the renaming. Default is TRUE.}

\item{...}{other arguments to be passed.}
}
\value{
an object of class `IFC_data`.
}
\description{
Renames populations in an `IFC_data` object
}
