% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings_utils.R
\name{random_name}
\alias{random_name}
\title{Random Name Generator}
\usage{
random_name(
  n = 10,
  ALPHA = LETTERS,
  alpha = letters,
  num = 0L:9L,
  special = c("#", "@", "?", "!", "&", "\%", "$"),
  forbidden = character()
)
}
\arguments{
\item{n}{number of characters of the desired return name. Default is 10.}

\item{ALPHA}{upper case letters. Default is LETTERS.}

\item{alpha}{lower case letters. Default is letters.}

\item{num}{integer to use. Default is 0:9}

\item{special}{characters. Default is c("#", "@", "?", "!", "&", "\%", "$").}

\item{forbidden}{forbidden character vector. Default is character().}
}
\value{
a character string.
}
\description{
Generates random name
}
\details{
'forbidden' should not encompass all possible returned value otherwise the function will never end.
}
\keyword{internal}
