% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs.R
\name{readFCS}
\alias{readFCS}
\title{FCS File Parser}
\source{
FCS specifications available at \url{http://murphylab.web.cmu.edu/FCSAPI/FCS3.html}.
}
\usage{
readFCS(
  fileName,
  options = list(header = list(version = list(at = 0, n = 6), text_beg = list(at = 10,
    n = 8), text_end = list(at = 18, n = 8), data_beg = list(at = 26, n = 8), data_end =
    list(at = 34, n = 8)), apply_scale = TRUE, first_only = TRUE),
  display_progress = TRUE,
  ...
)
}
\arguments{
\item{fileName}{path to file.}

\item{options}{list of options used to parse FCS file. It should contain:\cr
- header, a list whose members define at the position where to read info and n the number of bytes to extract:\cr
-- version: where to retrieve file version. Default is list(at = 0, n = 6),\cr
-- text_beg: where to retrieve file text segment beginning. Default is list(at = 10, n = 8),\cr
-- text_end: where to retrieve file text segment end.       Default is list(at = 18, n = 8),\cr
-- data_beg: where to retrieve file text segment beginning. Default is list(at = 26, n = 8),\cr
-- data_end: where to retrieve file text segment beginning. Default is list(at = 34, n = 8),\cr
- apply_scale, whether to apply data scaling. Default is TRUE\cr
- first_only, whether to extract only first. Default is FALSE}

\item{display_progress}{whether to display a progress bar. Default is TRUE.}

\item{...}{other arguments to be passed.}
}
\value{
a list whose elements are lists of each dataset stored within the file.\cr
each sublist contains:\cr
- header, list of header information corresponding to 'options'\cr
- delimiter, unique character used to separate keyword / values\cr
- text, list of keywords values,\cr
- data, data.frame of values.
}
\description{
Parse data from Flow Cytometry Standard compliant files.
}
\details{
'options' may be tweaked according to file type, instrument and software used to generate it. Default 'options' should allow to read most files.
}
