% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIFC.R
\name{readIFC}
\alias{readIFC}
\title{IFC Files Generic Reader}
\usage{
readIFC(fileName, ...)
}
\arguments{
\item{fileName}{path to file.}

\item{...}{arguments to pass to \code{\link{ExtractFromDAF}} or \code{\link{ExtractFromXIF}}.}
}
\description{
Extracts IFC data from IFC files no matter if they are DAF, RIF or CIF.
}
\details{
see \code{\link{ExtractFromDAF}} or \code{\link{ExtractFromXIF}}.
}
\examples{
if(requireNamespace("IFCdata", quietly = TRUE)) {
  ## use a rif file, but you can also read daf or cif
  file_rif <- system.file("extdata", "example.rif", package = "IFCdata")
  rif <- readIFC(fileName = file_rif)
} else {
  message(sprintf('Please run `install.packages("IFCdata", repos = "\%s", type = "source")` \%s',
                  'https://gitdemont.github.io/IFCdata/',
                  'to install extra files required to run this example.'))
}
}
