% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redefine_utils.R
\name{switch_channel}
\alias{switch_channel}
\title{IFC_data Channel Swap}
\usage{
switch_channel(obj, from, to)
}
\arguments{
\item{obj}{an `IFC_data` object. Default is missing.}

\item{from, to}{integer, determining channels to swap}

\item{...}{Other arguments to be passed.}
}
\value{
an `IFC_data` object.
}
\description{
Helper to swap channels in a `IFC_data` object.
}
\details{
it allows to apply all computation based on channel 'from' onto channel 'to', and reversely
}
\keyword{internal}
