% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toBIN_features.R
\name{toBIN_features}
\alias{toBIN_features}
\title{IFC_features Raw Conversion}
\usage{
toBIN_features(
  features,
  endianness = .Platform$endian,
  verbose = FALSE,
  display_progress = TRUE,
  pb_title = ""
)
}
\arguments{
\item{features}{an IFC_features object.}

\item{verbose}{whether to display message about current action. Default is FALSE.}

\item{display_progress}{whether to display a progress bar. Default is TRUE.}

\item{pb_title}{character string, giving the window title on the dialog box. Default is "".}
}
\value{
an raw vector of features binaries.
}
\description{
Helper to convert features (IFC_features object) to raw vector.
}
\keyword{internal}
