% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoami.R
\name{whoami}
\alias{whoami}
\title{Input Identification}
\usage{
whoami(
  entries = as.list(match.call()),
  search = list(info = "IFC_info", param = "IFC_param", offsets = "IFC_offset"),
  reinit = TRUE
)
}
\arguments{
\item{entries}{arguments from the function \code{\link{whoami}} is called.
/!\ \code{\link{whoami}} MUST be called explicitly this way: whoami(entries = as.list(match.call())).}

\item{search}{a named list of classes to search for entries.}

\item{reinit}{whether to reinitialize arguments to their default values in called environment. Default is TRUE.}
}
\value{
a list whose members are 'fileName': value of fileName if provided as a named argument 
in entries and all classes defined in 'search'
}
\description{
Helper that identifies input arguments thanks to their IFC classes even if they are not or mis named.
}
\keyword{internal}
