% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoami.R
\name{whoami}
\alias{whoami}
\title{Input Identification}
\usage{
whoami(entries)
}
\arguments{
\item{entries}{arguments from the function \code{\link{whoami}} is called.}
}
\value{
a list whose members are:
-fileName: value of fileName if provided as a named argument,\cr
-info; value of `IFC_info` object identified,\cr
-param: value of `IFC_param` object identified,\cr
-offsets: value of `IFC_offset` object identified,\cr
-ifd: value of `IFC_ifd` object identified,\cr
-data: value of `IFC_data` object identified.\cr
}
\description{
Helper that identifies input arguments thanks to their IFC classes even if they are not named.
}
\keyword{internal}
